% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_any_miss}
\alias{add_any_miss}
\title{Add a column that tells you if there are any missing values}
\usage{
add_any_miss(data, ..., label = "any_miss")
}
\arguments{
\item{data}{data.frame}

\item{...}{Variable names to use instead of the whole dataset. By default this
looks at the whole dataset. Otherwise, this is one or more unquoted
expressions separated by commas. These also respect the dplyr verbs
"starts_with", "contains", "ends_with", etc. By default will add "_all" to
the label if left blank, otherwise will add "_vars" to distinguish that it
has not been used on all of the variables.}

\item{label}{label for the column, defaults to "any_miss"}
}
\value{
data.frame with data and the column labelling whether that row (for
    those variables) has any missing values - indicated by "missing" and
    "complete".
}
\description{
Add a column that tells you if there are any missing values
}
\note{
perhaps there should be a general function for this
}
\examples{

airquality \%>\% add_any_miss()
airquality \%>\% add_any_miss(Ozone, Solar.R)

}
