% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_label_missings}
\alias{add_label_missings}
\title{Add a column describing if there are any missings in the dataset}
\usage{
add_label_missings(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame with a column "any_missing" that is either "Not Missing" or "Missing" for the purposes of plotting / exploration / nice print methods
}
\description{
Add a column describing if there are any missings in the dataset
}
\examples{

airquality \%>\% add_label_missings()

}
