% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_label_shadow}
\alias{add_label_shadow}
\title{Add a column describing whether there is a shadow}
\usage{
add_label_shadow(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame with
}
\description{
Add a column describing whether there is a shadow
}
\examples{

airquality \%>\%
add_shadow(Ozone, Solar.R) \%>\%
add_label_shadow()

}
