% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{add_n_miss}
\alias{add_n_miss}
\title{Add column containing number of missing data values}
\usage{
add_n_miss(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
It can be useful when doing data analysis to add the number of missing data points into your dataframe. add_n_miss adds a column named "n_miss", which contains the number of missing values in that row.
}
\examples{
library(magrittr)
airquality \%>\% add_n_miss()


}
