% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-n-prop-miss.R
\name{add_n_na}
\alias{add_n_na}
\title{Add column containing number of missing data values}
\usage{
add_n_na(data, ..., label = "n_na")
}
\arguments{
\item{data}{a dataframe}

\item{...}{Variable names to use instead of the whole dataset. By default this
looks at the whole dataset. Otherwise, this is one or more unquoted
expressions separated by commas. These also respect the dplyr verbs
\code{starts_with}, \code{contains}, \code{ends_with}, etc. By default will add "_all" to
the label if left blank, otherwise will add "_vars" to distinguish that it
has not been used on all of the variables.}

\item{label}{character default is "n_na".}
}
\value{
a dataframe
}
\description{
It can be useful when doing data analysis to add the number of missing data
points into your dataframe. \code{add_n_na} adds a column named "n_na",
which contains the number of missing values in that row.
}
\examples{

airquality \%>\% add_n_na()
airquality \%>\% add_n_na(Ozone, Solar.R)
airquality \%>\% add_n_na(dplyr::contains("o"))


}
