% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_time_series.R
\name{add_period_counter}
\alias{add_period_counter}
\title{Add a counter variable for the period of time}
\usage{
add_period_counter(data, period_length)
}
\arguments{
\item{data}{data.frame}

\item{period_length}{integer}
}
\value{
data.frame with extra variable "period_counter".
}
\description{
This adds a period_counter variable to the dataframe. It is used internally
inside `narnia` to facilitate the counting of missing values over a given
time period.
}
\examples{
\dontrun{
library(imputeTS)
add_period_counter(as_tibble(tsNH4),
                   period_length = 3)
}

}
