% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_shadow}
\alias{add_shadow}
\title{Add a column of the shadows to the dataframe}
\usage{
add_shadow(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{One or more unquoted expressions separated by commas. These also
respect the dplyr verbs "starts_with", "contains", "ends_with", etc.}
}
\value{
data.frame
}
\description{
Add a column of the shadows to the dataframe
}
\examples{

airquality \%>\% add_shadow(Ozone)
airquality \%>\% add_shadow(Ozone, Solar.R)

}
