% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_shadow_shift}
\alias{add_shadow_shift}
\title{Add a shadow shifted column to a dataset}
\usage{
add_shadow_shift(data, vars, suffix = "shift")
}
\arguments{
\item{data}{data.frame or .tbl}

\item{vars}{quoted variables}

\item{suffix}{suffix to add to variable, defaults to "shift"}
}
\value{
.data with the added variable shifted named as `var_suffix`
}
\description{
Shifting the values of a numeric
}
\examples{

add_shadow_shift(pedestrian, vars = "hourly_counts")
add_shadow_shift(airquality, vars = c("Ozone", "Solar.R"))

}
