% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_span_counter}
\alias{add_span_counter}
\title{Add a counter variable for a span of dataframe}
\usage{
add_span_counter(data, span_size)
}
\arguments{
\item{data}{data.frame}

\item{span_size}{integer}
}
\value{
data.frame with extra variable "span_counter".
}
\description{
Adds a variable, `span_counter` to a dataframe. Used internally to facilitate
counting of missing values over a given span.
}
\examples{
\dontrun{
add_span_counter(pedestrian, span_size = 100)
}
}
