% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all-is-miss-complete}
\alias{all-is-miss-complete}
\alias{all_na}
\alias{all_miss}
\alias{all_complete}
\title{Identify if all values are missing or complete}
\usage{
all_na(x)

all_miss(x)

all_complete(x)
}
\arguments{
\item{x}{an R object to be tested.}
}
\description{
This is shorthand for \code{all(is.na(x))} and \code{all(!is.na(x))}
}
\examples{

misses <- c(NA, NA, NA)
complete <- c(1, 2, 3)
mixture <- c(NA, 1, NA)

all_na(misses)
all_na(complete)
all_na(mixture)
all_complete(misses)
all_complete(complete)
all_complete(mixture)

}
