% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_row_complete}
\alias{all_row_complete}
\title{Helper function to determine whether all rows are complete}
\usage{
all_row_complete(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical vector
}
\description{
Helper function to determine whether all rows are complete
}
