% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_row_na}
\alias{all_row_na}
\title{Helper function to determine whether all rows are missing}
\usage{
all_row_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical vector
}
\description{
Helper function to determine whether all rows are missing
}
