% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{any-na}
\alias{any-na}
\alias{any_na}
\alias{any_miss}
\alias{any_complete}
\title{Identify if there are any missing or complete values}
\usage{
any_na(x)

any_miss(x)

any_complete(x)
}
\arguments{
\item{x}{an R object to be tested}
}
\description{
It is useful to search for any instances of missing or complete values. There
Are two functions that do this in \code{naniar} - \code{any_miss} and it's alias
\code{any_na}. These bother under the hood call \code{anyNA}. \code{any_complete} is
the complement to \code{any_miss} - it returns TRUE if there are any complete values.
}
\examples{

anyNA(airquality)
any_na(airquality)
any_miss(airquality)
any_complete(airquality)


}
\seealso{
\code{\link[=all_miss]{all_miss()}} \link{all_complete}
}
