% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{any_na}
\alias{any_na}
\title{any_na}
\usage{
any_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Logical TRUE / FALSE
}
\description{
find whether a vector contains a missing value. Used internally.
}
\examples{

library(ggmissing)
any_na(airqality$Solar.R)
any_na(airqality$Wind)

}

