% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any_row_na}
\alias{any_row_na}
\title{Helper function to determine whether there are any missings}
\usage{
any_row_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical vector TRUE = missing FALSE = complete
}
\description{
Helper function to determine whether there are any missings
}
