% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{are_shadow}
\alias{are_shadow}
\title{Are these things shadows?}
\usage{
are_shadow(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
logical vector - TRUE if contains a variable with a column ending in "_NA"
}
\description{
Does this thing contain a shadow variable?
}
\examples{

df_shadow <- bind_shadow(airquality)

are_shadow(df_shadow)

}
