% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{as_shadow.data.frame}
\alias{as_shadow.data.frame}
\title{Create shadow data}
\usage{
\method{as_shadow}{data.frame}(data, ...)
}
\arguments{
\item{data}{dataframe}

\item{...}{selected variables to use}
}
\description{
Return a tibble that in shadow matrix form, where the variables are the same but have a suffix _NA attached to indicate their difference.
}
\examples{

as_shadow(airquality)

}
