% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{as_shadow}
\alias{as_shadow}
\title{S3 method for create shadows}
\usage{
as_shadow(data, ...)
}
\arguments{
\item{data}{dataframe}

\item{...}{selected variables to use}
}
\value{
appended shadow with column names
}
\description{
Shadows have _NA as a suffix
}
