% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{as_shadow_upset}
\alias{as_shadow_upset}
\title{Convert data into shadow format for doing an upset plot}
\usage{
as_shadow_upset(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame
}
\description{
Upset plots are a way of visualising common sets, this function transforms
the data into a format that feeds directly into an upset plot
}
\examples{

\dontrun{

library(UpSetR)
airquality \%>\%
  as_shadow_upset() \%>\%
  upset()
}

}
