% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{bind_shadow}
\alias{bind_shadow}
\title{Column bind a shadow dataframe to original data}
\usage{
bind_shadow(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
dataframe with an extra appended layer of data.
}
\description{
Binding a shadow dataframe to a regular dataframe helps visualise and work with missing data
}
\examples{

bind_shadow(airquality)

# explore missing data visually
library(ggplot2)

# using the bounded shadow

ggplot(data = bind_shadow(airquality),
       aes(x = Ozone)) +
       geom_histogram() +
       facet_wrap(~Solar.R_NA,
       ncol = 1)

# using the is_na function

ggplot(data = airquality,
       aes(x = Ozone)) +
  geom_histogram() +
  facet_wrap(~is_na(Solar.R),
             ncol = 1)

}
