% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-breizh.R
\docType{data}
\name{breizh}
\alias{breizh}
\title{Data in Open Data, from the Breizh Open Data Plateform
2016.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1209 rows and 37 columns.}
\source{
\url{https://breizh-sba.opendatasoft.com/explore/dataset/marches-publics-conseil-regional-de-bretagne/information/}
}
\usage{
data(breizh)
}
\description{
The data is a subset of a dataset avalaible on the Breizh Open Data
Plateform. It contains the procurement contracts made by the
Conseil Regional de Bretagne from 2013 to 2015.
}
\details{
Full dataset is available here:
\url{https://breizh-sba.opendatasoft.com/explore/dataset/marches-publics-conseil-regional-de-bretagne/information/}

Format: a data frame with 1209 observations of 37 variables.
}
\keyword{datasets}
