% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{cast_shadow_shift_label}
\alias{cast_shadow_shift_label}
\title{Add a shadow and a shadow_shift column to a dataset}
\usage{
cast_shadow_shift_label(data, vars)
}
\arguments{
\item{data}{data.frame}

\item{vars}{character string for variables}
}
\value{
data.frame with the shadow and shadow_shift vars, and missing labels
}
\description{
Shift the values, add shadow, add missing label
}
\examples{

airquality \%>\% cast_shadow_shift_label("Ozone")
airquality \%>\% cast_shadow_shift_label(c("Ozone", "Solar.R"))

}
