% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_missing.R
\name{col_missing}
\alias{col_missing}
\title{col_missing}
\usage{
col_missing(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
Return a tidy dataframe that contains the number of missings in each column
}
\examples{

col_missing(airquality)

}

