% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-common-na-strings.R
\docType{data}
\name{common_na_strings}
\alias{common_na_strings}
\title{Common string values for NA}
\format{An object of class \code{character} of length 18.}
\usage{
common_na_strings
}
\description{
This vector contains common values of NA (missing), which is aimed to
be used inside naniar functions \code{\link[=miss_scan_count]{miss_scan_count()}} and
}
\examples{

dat_ms <- tibble::tribble(~x,  ~y,    ~z,
                          1,   "A",   -100,
                          3,   "N/A", -99,
                          NA,  NA,    -98,
                          -99, "E",   -101,
                          -98, "F",   -1)

miss_scan_count(dat_ms, -99)
miss_scan_count(dat_ms, c(-99,-98))
miss_scan_count(dat_ms, c("-99","-98","N/A"))
miss_scan_count(dat_ms, common_na_strings)
}
\keyword{datasets}
