% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_missing_point}
\title{Key drawing functions}
\usage{
draw_key_missing_point(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Each Geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These are the options built into naniar.
}
\keyword{internal}
