% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{gather_shadow}
\alias{gather_shadow}
\title{Long form representation of a shadow matrix}
\usage{
gather_shadow(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
dataframe in long, format, containing information about the missings
}
\description{
`gather_shadow` is a long-form representation of binding the shadow matrix to
  your data, producing variables named `rows`, `var`, and `miss`, where
  `miss` contains the missing value representation.
}
\examples{

gather_shadow(airquality)

}
