% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing_histogram.R
\name{geom_missing_histogram}
\alias{geom_missing_histogram}
\title{geom_missing_histogram}
\usage{
geom_missing_histogram(mapping = NULL, data = NULL, position = "stack",
  colour = ..missing.., binwidth = NULL, bins = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{?}{unknown parameters...}
}
\description{
geom_missing_histogram adds a point geometry for displaying missingness.
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.

Data plotting works. Still todo:
manipulate the colour aes so that the colours and legend appear.
fix awful default point sizes.
Warning message if na.rm = T is supplied.
}

