% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing_point.R
\name{geom_missing_point}
\alias{geom_missing_point}
\title{geom_missing_point}
\usage{
geom_missing_point(mapping = NULL, data = NULL, position = "identity",
  colour = ..missing.., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{other arguments passed on to \code{layer}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
geom_missing_point adds a point geometry for displaying missingness.
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
manipulate the colour aes so that the colours and legend appear.
fix awful default point sizes.
Warning message if na.rm = T is supplied.
}
\examples{

library(ggplot2)
library(naniar)

# using regular geom_point()
ggplot(airquality,
       aes(x = Ozone,
           y = Solar.R)) +
geom_point()

# using geom_missing_point()
ggplot(airquality,
       aes(x = Ozone,
           y = Solar.R)) +
geom_missing_point()

}

