% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing.R
\docType{data}
\name{GeomMissing}
\alias{GeomMissing}
\title{geom_missing}
\format{An object of class \code{GeomMissing} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomMissing
}
\arguments{
\item{?}{unknown parameters...}
}
\description{
geom_missing adds a geometry for displaying missingness.
}
\details{
\code{geom_missing} does this thing
}
\note{
At this stage geom_missing is limited to geom_point

Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
manipulate the colour aes so that the colours and legend appear.
fix awful default point sizes.
Warning message if na.rm = T is supplied.

ggplot2 must be installed from GitHub (devtools::install_github("hadley/ggplot2")) to obtain the function `ggproto`
}
\keyword{datasets}

