% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing_point.R
\docType{data}
\name{GeomMissingPoint}
\alias{GeomMissingPoint}
\title{geom_missing_point}
\format{An object of class \code{GeomMissingPoint} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomMissingPoint
}
\arguments{
\item{?}{unknown parameters...}
}
\description{
geom_missing_point adds a point geometry for displaying missingness.
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
manipulate the colour aes so that the colours and legend appear.
fix awful default point sizes.
Warning message if na.rm = T is supplied.
}
\keyword{datasets}

