% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-case-cumsum.R
\name{gg_miss_case_cumsum}
\alias{gg_miss_case_cumsum}
\title{Plot of cumulative sum of missing for cases}
\usage{
gg_miss_case_cumsum(x, breaks = 20)
}
\arguments{
\item{x}{a dataframe}

\item{breaks}{the breaks for the x axis default is 20}
}
\value{
a ggplot object depicting the number of missings
}
\description{
A plot showing the cumulative sum of missing values for cases, reading the
rows from the top to bottom. A default minimal theme is used, which can be
customised as normal for ggplot.
}
\examples{

gg_miss_case_cumsum(airquality)
}
\seealso{
\code{\link[=geom_miss_point]{geom_miss_point()}} \code{\link[=gg_miss_case]{gg_miss_case()}} \code{\link[=gg_miss_fct]{gg_miss_fct()}} \code{\link[=gg_miss_span]{gg_miss_span()}} \code{\link[=gg_miss_var]{gg_miss_var()}} \code{\link[=gg_miss_var_cumsum]{gg_miss_var_cumsum()}} \code{\link[=gg_miss_which]{gg_miss_which()}}
}
