% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_fac}
\alias{gg_miss_fac}
\title{Plot the number of missings for each variable, broken down by a factor}
\usage{
gg_miss_fac(x, fac)
}
\arguments{
\item{x}{a dataframe}

\item{fac}{the column containing the factor variable}
}
\value{
a ggplot object depicting the number of missings
}
\description{
This function draws a ggplot plot of the number of missings in each column, broken down by a categorical variable from the dataset.
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

gg_miss_fac(x = riskfactors, fac = marital)
library(ggplot2)
gg_miss_fac(x = riskfactors, fac = marital) + theme_bw()
gg_miss_fac(x = riskfactors, fac = marital) + labs(title = "NA in Risk Factors and Marital status")
}
