% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_time_series.R
\name{gg_miss_ts}
\alias{gg_miss_ts}
\title{Plot the number of missings in a given period}
\usage{
gg_miss_ts(dat_ts, period)
}
\arguments{
\item{dat_ts}{ts object}

\item{period}{integer describing the length of the period to be explored}
}
\value{
ggplot2 object
}
\description{
This is a replacement function to
imputeTS::plotNA.distributionBar(tsNH4, breaksize = 100), which shows the
number of missings in a given time period
}
\examples{
\dontrun{
library(imputeTS)
gg_miss_ts(tsNH4, period = 100)
}
}
