% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-var-cumsum.R
\name{gg_miss_var_cumsum}
\alias{gg_miss_var_cumsum}
\title{Plot of cumulative sum of missing value for each variable}
\usage{
gg_miss_var_cumsum(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a ggplot object showing the cumulative sum of missings over the variables
}
\description{
A plot showing the cumulative sum of missing values for each variable,
reading columns from the left to the right of the initial dataframe. A
default minimal theme is used, which can be customised as normal for ggplot.
}
\examples{

gg_miss_var_cumsum(airquality)
}
\seealso{
\code{\link[=geom_miss_point]{geom_miss_point()}} \code{\link[=gg_miss_case]{gg_miss_case()}} \code{\link[=gg_miss_case_cumsum]{gg_miss_case_cumsum()}} \code{\link[=gg_miss_fct]{gg_miss_fct()}} \code{\link[=gg_miss_span]{gg_miss_span()}} \code{\link[=gg_miss_var]{gg_miss_var()}} \code{\link[=gg_miss_which]{gg_miss_which()}}
}
