% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_which}
\alias{gg_miss_which}
\title{Plot which variables contain a missing value}
\usage{
gg_miss_which(x, xlab = " ", ylab = " ", title = NULL, ...)
}
\arguments{
\item{x}{a dataframe}

\item{xlab}{legend for the x axis. Default is " ".}

\item{ylab}{legend for the y axis. Default is " ".}

\item{title}{title for the plot. Default is NULL.}

\item{...}{you can pass more ggplot theme parameters inside the ...}
}
\value{
a ggplot plot
}
\description{
`gg_miss_which` (need a better name!) produces a set of rectangles that indicate whether there is a missing element in a column or not
}
\examples{

gg_miss_which(airquality)
gg_miss_which(airquality, xlab = "Variables", ylab = "Height")
gg_miss_which(airquality, title = "Missing values in airquality")
library(ggplot2)
gg_miss_which(airquality, panel.background = element_rect(fill = "grey"))
}
