% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_missing_var.R
\name{gg_missing_var}
\alias{gg_missing_var}
\title{gg_missing_var}
\usage{
gg_missing_var(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot plot depicting the number of missings in a given column
}
\description{
This function draws a ggplot plot of the number of missings in each column, rearranged to show which variables have the most missing data. At this point I'm still debating whether or not this should be with `geom_bar` or with `geom_lollipop` from hrbrmstr's `ggalt` package.
}
\examples{

gg_missing_var(airquality)

}

