% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_row_missing.R
\name{gg_row_missing}
\alias{gg_row_missing}
\title{gg_row_missing}
\usage{
gg_row_missing(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot plot depicting the number of missings in a given row
}
\description{
This function draws a ggplot plot of the number of missings in each row.
At this point I'm not sure how this plot should be arranged, as it currently looks ugly!
}
\examples{

gg_row_missings(airquality)

}

