% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_below.R
\name{impute_below_at}
\alias{impute_below_at}
\title{Scoped variants of \code{impute_below}}
\usage{
impute_below_at(.tbl, .vars, ...)
}
\arguments{
\item{.tbl}{a data.frame}

\item{.vars}{variables to impute}

\item{...}{extra arguments}
}
\value{
an dataset with values imputed
}
\description{
\code{impute_below} operates on all variables. To only impute variables
that satisfy a specific condition, use the scoped variants,
\code{impute_below_at}, and \code{impute_below_if}.
}
\examples{
# select variables starting with a particular string.
library(dplyr)
impute_below_at(airquality,
                .vars = starts_with("Oz"))

impute_below_at(airquality,
                .vars = 1:2)

impute_below_at(airquality,
                .vars = everything())


}
