% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_knn.R
\name{impute_knn}
\alias{impute_knn}
\title{Impute nearest neighbours}
\usage{
impute_knn(.tbl, k = 5, method = "mean")
}
\arguments{
\item{.tbl}{data.frame with missing values}

\item{k}{integer, default value 5}

\item{method}{\itemize{
\item "mean" or "random" - do you want to take the mean of the k
neighbours, do you want to take a random value from the k neighbours.
}}
}
\value{
data.frame with imputed values
}
\description{
Impute nearest neighbours
}
\note{
need to add separate functions for knn_avg and knn_rand, and also
scoped variants. Ideally, this should work on a vector first, with
different methods for each part
}
\examples{
airquality \%>\%
  bind_shadow() \%>\%
  impute_knn("mean")

\dontrun{
library(ggplot2)
airquality \%>\%
  bind_shadow() \%>\%
  impute_knn() \%>\%
  add_label_shadow() \%>\%
  ggplot(aes(x = Ozone,
             y = Solar.R,
             colour = any_missing)) +
         geom_point()
}
}
