% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-median.R
\name{impute_median}
\alias{impute_median}
\alias{impute_median.default}
\alias{impute_median.factor}
\title{Impute the median value into a vector with missing values}
\usage{
impute_median(x)

\method{impute_median}{default}(x)

\method{impute_median}{factor}(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with median values replaced
}
\description{
Impute the median value into a vector with missing values
}
\examples{

vec <- rnorm(10)

vec[sample(1:10, 3)] <- NA

impute_median(vec)

}
