% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_shift.R
\name{impute_shift_at}
\alias{impute_shift_at}
\title{Scoped variants of \code{impute_shift}}
\usage{
impute_shift_at(.tbl, .vars, ...)
}
\arguments{
\item{.tbl}{a data.frame}

\item{.vars}{variables to impute}

\item{...}{extra arguments}
}
\value{
an dataset with values imputed
}
\description{
\code{impute_shift} operates on all variables. To only impute variables
that satisfy a specific condition, use the scoped variants,
\code{impute_shift_at}, and \code{impute_shift_if}.
}
\examples{
# select variables starting with a particular string.
library(dplyr)
impute_shift_at(airquality,
                .vars = starts_with("Oz"))

impute_shift_at(airquality,
                .vars = 1:2)

impute_shift_at(airquality,
                .vars = everything())

# using the "_if" scoped variant


}
