% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_shift.R
\name{impute_shift_if}
\alias{impute_shift_if}
\title{Scoped variants of \code{impute_shift}}
\usage{
impute_shift_if(.tbl, .predicate, ...)
}
\arguments{
\item{.tbl}{data.frame}

\item{.predicate}{A predicate function (such as is.numeric)}

\item{...}{extra arguments}
}
\value{
an dataset with values imputed
}
\description{
\code{impute_shift} operates on all variables. To only impute variables
that satisfy a specific condition, use the scoped variants,
\code{impute_shift_at}, and \code{impute_shift_if}.
}
\examples{

airquality \%>\%
  impute_shift_if(.predicate = is.numeric)

}
