% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{is_na}
\alias{is_na}
\title{Give NAs a more meaningful label}
\usage{
is_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
Returns a binary factor of !NA and NA, where !NA indicates a datum that is not missing, and NA indicats missingness.
}
\examples{

is_na(airquality$Ozone)

}
\seealso{
as_shadow
}

