% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-verifiers.R
\name{is_shadow}
\alias{is_shadow}
\alias{are_shadow}
\title{Test if input is or are shadow variables}
\usage{
is_shadow(x)

are_shadow(x)
}
\arguments{
\item{x}{a vector or data.frame}
}
\value{
logical vector of length 1
}
\description{
Creating shadow variables can be done with \code{bind_shadow}. To see if there
are any shadow variables, \code{is_shadow} will do this.
}
\examples{

aq_sh <- as_shadow(airquality)
aq_bind <- bind_shadow(airquality)

is_shadow(aq_sh)
are_shadow(aq_sh)
is_shadow(aq_bind)
are_shadow(aq_bind)

}
