% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_missings.R
\name{label_missing_1d}
\alias{label_missing_1d}
\title{label_missing_1d}
\usage{
label_missing_1d(x1)
}
\arguments{
\item{x1}{a variable of a dataframe}
}
\value{
a vector indicating whether any of these rows had missing values
}
\description{
Label whether a value is missing in either row of two columns. At the moment this is a more appealing alternative to miss_cat, which is at this stage a bit complicated.
}
\note{
can we generalise label_missing to work for any number of variables?
}
\examples{

label_missing_1d(airquality$Ozone)

}

