% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-label.R
\name{label_missing_2d}
\alias{label_missing_2d}
\title{label_missing_2d}
\usage{
label_missing_2d(x1, x2)
}
\arguments{
\item{x1}{a variable of a dataframe}

\item{x2}{another variable of a dataframe}
}
\value{
a vector indicating whether any of these rows had missing values
}
\description{
Label whether a value is missing in either row of two columns. This is a more appealing alternative to miss_cat, which seems a bit overcomplicated.
}
\examples{

label_missing_2d(airquality$Ozone, airquality$Solar.R)

}
