% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-label.R
\name{label_missings}
\alias{label_missings}
\title{Is there a missing value in the row of a dataframe?}
\usage{
label_missings(data)
}
\arguments{
\item{data}{a dataframe or set of vectors of the same length}
}
\value{
character vector of "Missing" and "Not Missing".
}
\description{
Creates a character vector describing whether there are missing or not missing
  values
}
\note{
should be an add_missing_labels function so you can do `data %>% add_missing_labels()`
}
\examples{

label_missings(airquality)

library(dplyr)

airquality \%>\% mutate(is_missing = label_missings(airquality))

}
