% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-miss.R
\name{label_na}
\alias{label_na}
\title{Give NAs a more meaningful label}
\usage{
label_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of factors containing the labels "!NA" for Not missing and
"NA" for missing.
}
\description{
Returns a binary factor of !NA and NA, where !NA indicates a datum that is
not missing, and NA indicates missingness. This function is what powers the
factor levels in \code{as_shadow()}.
}
\examples{
\dontrun{
label_na(airquality$Ozone)
}
}
\seealso{
as_shadow
}
