% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-miss.R
\name{label_na_1d}
\alias{label_na_1d}
\title{Label a missing from one column}
\usage{
label_na_1d(x1)
}
\arguments{
\item{x1}{a variable of a dataframe}
}
\value{
a vector indicating whether any of these rows had missing values
}
\description{
Label whether a value is missing in a row of one columns.
}
\note{
can we generalise label_na to work for any number of variables?
}
\examples{
\dontrun{
label_na_1d(airquality$Ozone)
}
}
