% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-miss.R
\name{label_na_2d}
\alias{label_na_2d}
\title{label_na_2d}
\usage{
label_na_2d(x1, x2)
}
\arguments{
\item{x1}{a variable of a dataframe}

\item{x2}{another variable of a dataframe}
}
\value{
a vector indicating whether any of these rows had missing values
}
\description{
Label whether a value is missing in either row of two columns.
}
\examples{
\dontrun{
label_na_2d(airquality$Ozone, airquality$Solar.R)
}
}
