% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{label_shadow}
\alias{label_shadow}
\title{Label shadow values as missing or not missing}
\usage{
label_shadow(data, ..., missing = "Missing", complete = "Not Missing")
}
\arguments{
\item{data}{data.frame}

\item{...}{extra variable to label}

\item{missing}{character a label for when values are missing - defaults to "Missing"}

\item{complete}{character character a label for when values are complete - defaults to "Not Missing"}
}
\value{
"Missing" or "Not Missing"
}
\description{
Powers \code{add_label_shadow}. For the moment it is an internal function.
}
