% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcar-test.R
\name{mcar_test}
\alias{mcar_test}
\title{Little's missing completely at random (MCAR) test}
\usage{
mcar_test(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row and four columns:
\item{statistic}{Chi-squared statistic for Little's test}
\item{df}{Degrees of freedom used for chi-squared statistic}
\item{p.value}{P-value for the chi-squared statistic}
\item{missing.patterns}{Number of missing data patterns in the data}
}
\description{
Use Little's (1988) test statistic to assess if data is missing completely at random (MCAR).
}
\examples{
mcar_test(airquality)
mcar_test(oceanbuoys)

# If there are non-numeric columns, there will be a warning
mcar_test(riskfactors)

}
\references{
Little, Roderick J. A. 1988. "A Test of Missing Completely at Random for Multivariate Data with Missing Values." \emph{Journal of the American Statistical Association} 83 (404): 1198--1202. \url{https://doi.org/10.1080/01621459.1988.10478722}.
}
