% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-complete-x-pct-prop.R
\name{miss-complete-case-pct}
\alias{miss-complete-case-pct}
\alias{miss_case_pct}
\alias{complete_case_pct}
\title{Percentage of cases that contain a missing or complete values.}
\usage{
miss_case_pct(data)

complete_case_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing or complete
value
}
\description{
Deprecated, please see \code{\link[=miss_case_pct]{miss_case_pct()}} and \code{\link[=complete_case_pct]{complete_case_pct()}}.
}
\seealso{
\code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
