% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-complete-x-pct-prop.R
\name{miss-complete-case-prop}
\alias{miss-complete-case-prop}
\alias{miss_case_prop}
\alias{complete_case_prop}
\title{Proportion of cases that contain a missing or complete values.}
\usage{
miss_case_prop(data)

complete_case_prop(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the proportion of cases that contain a missing or complete
value
}
\description{
Calculate the proportion of cases (rows) that contain missing or complete
values.
}
\examples{

miss_case_prop(airquality)
complete_case_prop(airquality)

}
\seealso{
\url{miss_case_pct} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_pct} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
