% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-complete-x-pct-prop.R
\name{miss-complete-var-prop}
\alias{miss-complete-var-prop}
\alias{miss_var_prop}
\alias{complete_var_prop}
\title{Proportion of variables containing missings or complete values}
\usage{
miss_var_prop(data)

complete_var_prop(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the proportion of variables that contain missing or complete
data
}
\description{
Calculate the proportion of variables that contain a single missing or
complete values.
}
\examples{

miss_var_prop(riskfactors)
miss_var_prop(oceanbuoys)
complete_var_prop(riskfactors)
complete_var_prop(oceanbuoys)

}
\seealso{
\url{miss_case_pct} \url{miss_case_prop} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_pct} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
