% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-complete-x-pct-prop.R
\name{miss-pct-prop-defunct}
\alias{miss-pct-prop-defunct}
\alias{miss_var_prop}
\alias{complete_var_prop}
\alias{miss_var_pct}
\alias{complete_var_pct}
\alias{miss_case_prop}
\alias{complete_case_prop}
\alias{miss_case_pct}
\alias{complete_case_pct}
\title{Proportion of variables containing missings or complete values}
\usage{
miss_var_prop(...)

complete_var_prop(...)

miss_var_pct(...)

complete_var_pct(...)

miss_case_prop(...)

complete_case_prop(...)

miss_case_pct(...)

complete_case_pct(...)
}
\arguments{
\item{...}{arguments}
}
\description{
Defunct. Please see \code{\link[=prop_miss_var]{prop_miss_var()}}, \code{\link[=prop_complete_var]{prop_complete_var()}}, \code{\link[=pct_miss_var]{pct_miss_var()}}, \code{\link[=pct_complete_var]{pct_complete_var()}}, \code{\link[=prop_miss_case]{prop_miss_case()}}, \code{\link[=prop_complete_case]{prop_complete_case()}}, \code{\link[=pct_miss_case]{pct_miss_case()}}, \code{\link[=pct_complete_case]{pct_complete_case()}}.
}
