% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_case_cumsum}
\alias{miss_case_cumsum}
\title{Summarise the missingness in each case}
\usage{
miss_case_cumsum(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a data_frame of the percent of missing data in each case
}
\description{
Provide a data_frame containing each case (row), the number of missing values in each case,  and the percent of missing values in each case.
}
\examples{

miss_case_cumsum(airquality)
miss_case_cumsum(breizh)

}
