% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pct-prop.R
\name{miss_case_pct}
\alias{miss_case_pct}
\title{Percentage of cases that contain a missing values.}
\usage{
miss_case_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing value
}
\description{
Calculate the percentage of cases (rows) that contain a missing value.
}
\examples{

miss_case_pct(airquality)

}
\seealso{
\url{miss_case_prop} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_pct} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
