% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pattern.R
\name{miss_pattern}
\alias{miss_pattern}
\title{Calculate the pattern of missingness}
\usage{
miss_pattern(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame of missingness patterns
}
\description{
Calculate the pattern of missingness
}
\examples{

miss_pattern(airquality)
}
