% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_time_series.R
\name{miss_ts_run}
\alias{miss_ts_run}
\title{Return the number of missing or complete values in a single run}
\usage{
miss_ts_run(dat_ts)
}
\arguments{
\item{dat_ts}{a ts object}
}
\value{
dataframe with column names "run_length" and "is_na", which describe
    the length of the run, and whether that run describes a missing value.
}
\description{
In time series it can be useful to determine the number of missing values
    that occur in a single run. This function `miss_ts_run` returns a
    dataframe with the column names "run_length" and "is_na", which describe
    the length of the run, and whether that run describes a missing value
}
\examples{

\dontrun{
library(imputeTS)
library(ggplot2)
library(dplyr)

# explore the number of missings in a given run
miss_ts_run(tsNH4) \%>\%
filter(is_na) \%>\%
count(run_length) \%>\%
ggplot(aes(x = run_length,
           y = n)) +
     geom_col()

# look at the number of missing values and the run length of these.
miss_ts_run(tsNH4) \%>\%
  ggplot(aes(x = is_na,
             y = run_length)) +
     geom_boxplot()

}

}
