% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_time_series.R
\name{miss_ts_summary}
\alias{miss_ts_summary}
\title{Summarise the number of missings in a given period for a ts object}
\usage{
miss_ts_summary(dat_ts, period)
}
\arguments{
\item{dat_ts}{ts object}

\item{period}{integer describing the length of the period to be explored}
}
\value{
dataframe with variables `n_miss`, `n_complete`, `prop_miss`, and
    `prop_complete`, which describe the number, or proportion of missing or
    complete values within that given time period
}
\description{
To summarise the missing values in a time series object it can be useful to
    calculate the number of missing values in a given time period.
    `miss_ts_summary` takes a `ts` object and a `period` argument and
     returns a dataframe containing the number of missing values within each
     period.
}
\examples{

\dontrun{
library(imputeTS)
miss_ts_summary(data_ts = tsNH4,
period = 100)
}

}
