% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R,
%   R/missing_data_tidiers.R
\name{percent_missing_var}
\alias{percent_missing_var}
\alias{miss_var_pct}
\title{Percentage of variables containing missings}
\usage{
percent_missing_var(data)

miss_var_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing data
}
\description{
Percentage of variables containing missings

Calculate the percentage of variables that contain a single missing value.
}
\examples{

library(naniar)

miss_var_pct(airquality)

}
