% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pct-prop.R
\name{miss_var_prop}
\alias{miss_var_prop}
\title{Proportion of variables containing missings}
\usage{
miss_var_prop(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the proportion of variables that contain missing data
}
\description{
Calculate the proportion of variables that contain a single missing value.
}
\examples{

miss_var_prop(riskfactors)
miss_var_prop(oceanbuoys)

}
