% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-var-case-scan.R
\name{miss_var_scan}
\alias{miss_var_scan}
\title{Scan variables for particular possible missings}
\usage{
miss_var_scan(data, scan)
}
\arguments{
\item{data}{data.frame}

\item{scan}{list of values to scan}
}
\value{
summary data.frame of variables and how many scanned vars
}
\description{
Scan variables for particular possible missings
}
\examples{

dat_ms <- tribble(~x,  ~y,    ~z,
                  3,   "N/A",  -99,
                  -99, "E",    97,
                  -98, "F",    97)

miss_var_scan(dat_ms,
              list("N/A",
                   -99))

miss_var_scan(dat_ms,
              list("N/A",
                   -99,
                   -98))
}
