% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-summary.R
\name{miss_var_summary}
\alias{miss_var_summary}
\title{Summarise the missingness in each variable}
\usage{
miss_var_summary(data, order = FALSE, ...)
}
\arguments{
\item{data}{a data.frame}

\item{order}{a logical indicating whether or not to order the result by n_miss. TRUE orders from largest to smallest n_miss, and FALSE orders by order provided by the data.}

\item{...}{extra arguments}
}
\value{
a tibble of the percent of missing data in each variable
}
\description{
Provide a summary for each variable of the number and percent missings,
ordering by the most missings in each variable.
}
\examples{

miss_var_summary(airquality)
miss_var_summary(oceanbuoys, order = TRUE)

# works with group_by from dplyr
library(dplyr)
airquality \%>\%
  group_by(Month) \%>\%
  miss_var_summary()

}
