% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where-na.R
\name{miss_var_which}
\alias{miss_var_which}
\title{Which variables contain missing values?}
\usage{
miss_var_which(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
character vector of variable names
}
\description{
It can be helpful when writing other functions to just return the names
of the variables that contain missing values. \code{miss_var_which} returns a
vector of variable names that contain missings. It will return NULL when
there are no missings.
}
\examples{
miss_var_which(airquality)

miss_var_which(mtcars)

}
