% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-var-miss.R
\name{n-var-case-complete}
\alias{n-var-case-complete}
\alias{n_var_complete}
\alias{n_case_complete}
\title{The number of variables with complete values}
\usage{
n_var_complete(data)

n_case_complete(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
integer number of complete values
}
\description{
This function calculates the number of variables that contain a complete value
}
\examples{

# how many variables contain complete values?
n_var_complete(airquality)
n_case_complete(airquality)

}
\seealso{
\code{\link[=n_var_miss]{n_var_miss()}}
}
