% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-var-miss.R
\name{n-var-case-miss}
\alias{n-var-case-miss}
\alias{n_var_miss}
\alias{n_case_miss}
\title{The number of variables or cases with missing values}
\usage{
n_var_miss(data)

n_case_miss(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
integer, number of missings
}
\description{
This function calculates the number of variables or cases that contain a missing value
}
\examples{
# how many variables contain missing values?
n_var_miss(airquality)
n_case_miss(airquality)

}
\seealso{
\code{\link[=n_var_complete]{n_var_complete()}}
}
