% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete-rows.R
\name{n_complete_row}
\alias{n_complete_row}
\title{Return a vector of the number of complete values in each row}
\usage{
n_complete_row(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric vector of the number of complete values in each row
}
\description{
Substitute for \code{rowSums(!is.na(data))}
}
\examples{

n_complete_row(airquality)
n_complete_row(pedestrian)

}
