% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{n_miss}
\alias{n_miss}
\title{return the number of missing values}
\usage{
n_miss(x)
}
\arguments{
\item{x}{a vector}
}
\description{
substitute for \code{sum(is.na(data))}
}
\examples{

n_miss(airquality)
n_miss(airquality$Ozone)

}

