% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete-rows.R
\name{n_miss_row}
\alias{n_miss_row}
\title{Return a vector of the number of missing values in each row}
\usage{
n_miss_row(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric vector of the number of missing values in each row
}
\description{
Substitute for \code{rowSums(is.na(data))}
}
\examples{

n_miss_row(airquality)
n_miss_row(pedestrian)

}
