% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pct-prop.R
\name{na_case_pct}
\alias{na_case_pct}
\title{Percentage of cases that contain a missing values.}
\usage{
na_case_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing value
}
\description{
Calculate the percentage of cases (rows) that contain a missing value.
}
\examples{

na_case_pct(airquality)

}
