% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pct-prop.R
\name{na_case_prop}
\alias{na_case_prop}
\title{Proportion of cases that contain a missing values.}
\usage{
na_case_prop(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the proportion of cases that contain a missing value
}
\description{
Calculate the proportion of cases (rows) that contain a missing value.
}
\examples{

na_case_prop(airquality)

}
