% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-prop-pct-summary.R
\name{na_prop_summary}
\alias{na_prop_summary}
\title{Proportions of missings in data, variables, and cases.}
\usage{
na_prop_summary(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Return missing data info about the dataframe, the variables, and the cases.
Specifically, returning how many elements in a dataframe contain a missing
value, how many elements in a variable contain a missing value, and how many
elements in a case contain a missing.
}
\examples{

na_prop_summary(airquality)

}
