% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-summary.R
\name{na_summary}
\alias{na_summary}
\title{Collate summary measures from naniar into one tibble}
\usage{
na_summary(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a tibble of missing data summaries
}
\description{
\code{na_summary} performs all of the missing data helper summaries and puts
them into lists within a tibble
}
\examples{

s_miss <- na_summary(airquality)
s_miss$na_df_prop
s_miss$na_case_table
# etc, etc, etc.

library(dplyr)
s_na_group <- group_by(airquality, Month) \%>\% na_summary()
s_na_group$na_df_prop
s_na_group$na_case_table
# etc, etc, etc.


}
