% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-cumsum.R
\name{na_var_cumsum}
\alias{na_var_cumsum}
\title{Cumulative sum of the number of missings in each variable}
\usage{
na_var_cumsum(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a tibble of the cumulative sum of missing data in each variable
}
\description{
Provide a data.frame containing the cumulative sum of number & percentage of
missingness for each variable.
}
\examples{

na_var_cumsum(airquality)

library(dplyr)

airquality \%>\%
  group_by(Month) \%>\%
  na_var_cumsum()

}
