% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-miss-point.R, R/naniar-ggproto.R,
%   R/stat-miss-point.R
\docType{data}
\name{GeomNaPoint}
\alias{GeomNaPoint}
\alias{naniar-ggproto}
\alias{StatNaPoint}
\title{naniar-ggroto}
\format{An object of class \code{StatNaPoint} (inherits from \code{Stat}, \code{ggproto}) of length 6.}
\usage{
StatNaPoint
}
\description{
These are the stat and geom overrides using ggproto from ggplot2
that make naniar work.
}
\keyword{datasets}
