% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naniar-package.R
\docType{package}
\name{naniar}
\alias{naniar}
\alias{naniar-package}
\title{naniar}
\description{
naniar is a package to make it easier to summarise and handle missing values
in R. It strives to do this in a way that is as consistent with tidyverse
principles as possible.
}
