% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-missing-point.R, R/narnia-ggproto.R,
%   R/stat-missing-point.R
\docType{data}
\name{GeomMissingPoint}
\alias{GeomMissingPoint}
\alias{narnia-ggproto}
\alias{StatMissingPoint}
\title{narnia-ggroto}
\format{An object of class \code{StatMissingPoint} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
StatMissingPoint
}
\description{
These are the stat and geom overrides using ggproto from ggplot2 that make narnia work.
}
\keyword{datasets}
