% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnia-package.R
\docType{package}
\name{narnia}
\alias{narnia}
\alias{narnia-package}
\title{narnia}
\description{
narnia is a package to make it easier to summarise and handle missing values
in R. It strives to do this in a way that is as consistent with tidyverse
principles as possible.
}
