% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabular.R
\name{new_nabular}
\alias{new_nabular}
\title{Create a new nabular format}
\usage{
new_nabular(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
object with class "nabular", inheriting from it's original class
}
\description{
Create a new nabular format
}
