% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{new_shadow}
\alias{new_shadow}
\title{Create a new shadow}
\usage{
new_shadow(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
object with class "shadow", inhereting from it's original class
}
\description{
Create a new shadow
}
