% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-oceanbuoys.R
\docType{data}
\name{oceanbuoys}
\alias{oceanbuoys}
\title{West Pacific Tropical Atmosphere Ocean Data, 1993 & 1997.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 736 rows and 8 columns.
}
\source{
\url{https://www.pmel.noaa.gov/tao/drupal/disdel/}
}
\usage{
data(oceanbuoys)
}
\description{

}
\details{
Format: a data frame with 736 observations on the following 8
variables.
\describe{
\item{\code{year}}{A numeric with levels \code{1993} \code{1997}.}
\item{\code{latitude}}{A numeric with levels \code{-5}  \code{-2} \code{0}.}
\item{\code{longitude}}{A numeric with levels \code{-110} \code{-95}.}
\item{\code{sea_temp_c}}{Sea surface temperature(degree Celsius),  measured
by the TAO buoys at one meter below the surface.}
\item{\code{air_temp_c}}{Air temperature(degree Celsius), measured by the
TAO buoys three meters above the sea surface.}
\item{\code{humidity}}{Relative humidity(\%), measured by the TAO buoys 3
meters above the sea surface.}
\item{\code{wind_ew}}{The East-West wind vector components(M/s).  TAO buoys
measure the wind speed and direction four meters above the sea surface.
If it is positive, the East-West component of the wind is blowing towards
the East. If it is negative, this component is blowing towards the West.}
\item{\code{wind_ns}}{The North-South wind vector components(M/s). TAO
buoys measure the wind speed and direction four meters above the sea
surface. If it is positive, the North-South component of the wind is
blowing towards the North. If it is negative, this component is blowing
towards the South.}}
}
\examples{

vis_miss(oceanbuoys)

# Look at the missingness in the variables
miss_var_summary(oceanbuoys)
\dontrun{
# Look at the missingness in air temperature and humidity
library(ggplot2)
p <-
ggplot(oceanbuoys,
       aes(x = air_temp_c,
           y = humidity)) +
     geom_miss_point()

 p

 # for each year?
 p + facet_wrap(~year)

 # this shows that there are more missing values in humidity in 1993, and
 # more air temperature missing values in 1997

 # see more examples in the vignette, "getting started with naniar".
}
}
\seealso{
library(MissingDataGUI) (data named "tao")
}
\keyword{datasets}
