% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop-pct-var-case-miss-complete.R
\name{pct-miss-complete-case}
\alias{pct-miss-complete-case}
\alias{pct_miss_case}
\alias{pct_complete_case}
\title{Percentage of cases that contain a missing or complete values.}
\usage{
pct_miss_case(data)

pct_complete_case(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing or complete
value
}
\description{
Calculate the percentage of cases (rows) that contain a missing or complete
value.
}
\examples{

pct_miss_case(airquality)
pct_complete_case(airquality)

}
\seealso{
\code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \code{\link[=miss_case_summary]{miss_case_summary()}} \code{\link[=miss_case_table]{miss_case_table()}} \code{\link[=miss_summary]{miss_summary()}} \code{\link[=miss_var_prop]{miss_var_prop()}} \code{\link[=miss_var_run]{miss_var_run()}} \code{\link[=miss_var_span]{miss_var_span()}} \code{\link[=miss_var_summary]{miss_var_summary()}} \code{\link[=miss_var_table]{miss_var_table()}}
}
