% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop-pct-var-case-miss-complete.R
\name{pct-miss-complete-var}
\alias{pct-miss-complete-var}
\alias{pct_miss_var}
\alias{pct_complete_var}
\title{Percentage of variables containing missings or complete values}
\usage{
pct_miss_var(data)

pct_complete_var(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing or complete
data
}
\description{
Calculate the percentage of variables that contain a single missing or
complete value.
}
\examples{

prop_miss_var(riskfactors)
prop_miss_var(oceanbuoys)
prop_complete_var(riskfactors)
prop_complete_var(oceanbuoys)

}
\seealso{
\code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
