% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete.R
\name{pct_complete}
\alias{pct_complete}
\title{Return the percent of complete values}
\usage{
pct_complete(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
numeric percent of complete values
}
\description{
The complement to `pct_miss`
}
\examples{

pct_complete(airquality)
pct_complete(airquality$Ozone)

}
