% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete.R
\name{pct_miss}
\alias{pct_miss}
\title{Return the percent of missing values}
\usage{
pct_miss(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
numeric the percent of missing values in x
}
\description{
This is shorthand for \code{mean(is.na(x)) * 100}
}
\examples{

pct_miss(airquality)
pct_miss(airquality$Ozone)

}
