% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete.R
\name{pct_na}
\alias{pct_na}
\title{Return the percent of missing values}
\usage{
pct_na(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
numeric the percent of missing values in x
}
\description{
This is shorthand for \code{mean(is.na(x)) * 100}
}
\examples{

pct_na(airquality)
pct_na(airquality$Ozone)

}
