% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R
\name{percent_missing_case}
\alias{percent_missing_case}
\title{Percentage of cases that contain a missing values.}
\usage{
percent_missing_case(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_case_pct instead.
}
\seealso{
miss_case_pct
}
