% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{percent_missing_var}
\alias{percent_missing_var}
\title{percent_missing_var}
\usage{
percent_missing_var(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
numeric the percent of variables that contain missing data
}
\description{
percent_missing_var
}
\examples{

library(naniar)

percent_missing_var(airquality)

}

