% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete-rows.R
\name{prop_miss_row}
\alias{prop_miss_row}
\title{Return a vector of the proportion of missing values in each row}
\usage{
prop_miss_row(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric vector of the proportion of missing values in each row
}
\description{
Substitute for \code{rowMeans(is.na(data))}
}
\examples{

prop_miss_row(airquality)
prop_miss_row(pedestrian)

}
