% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete.R
\name{prop_na}
\alias{prop_na}
\title{Return the proportion of missing values}
\usage{
prop_na(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
numeric the proportion of missing values in x
}
\description{
This is shorthand for \code{mean(is.na(x))}
}
\examples{

prop_na(airquality)
prop_na(airquality$Ozone)

}
