% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R, R/naniar-package.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{vis_miss}
\alias{\%>\%}
\alias{is_na}
\alias{are_na}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{is_na}}, \code{\link[rlang]{are_na}}}

  \item{visdat}{\code{\link[visdat]{vis_miss}}}
}}

