% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_with.R
\name{replace_na_with}
\alias{replace_na_with}
\title{Replace NA value with provided value}
\usage{
replace_na_with(..., .pytype = NULL, .size = NULL)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>

One or more vectors. These will be
\link[vctrs:vector_recycling_rules]{recycled} against each other, and will be
cast to their common type.}

\item{.size}{An optional size declaring the desired output size. If supplied,
this overrides the common size of the vectors in \code{...}.}
}
\description{
This function helps you replace NA values with a single provided value.
This can be classed as a kind of imputation, and so is similar to
\code{\link[=impute_fixed]{impute_fixed()}}. This function is actually a wrapper around
\code{\link[dplyr:coalesce]{dplyr::coalesce()}}. We renamed it to better evoke the use of replacing NA
values with a provided value, and expand the documentation to demonstrate
more use cases.
}
\examples{

library(naniar)
x <- sample(c(1:5, NA, NA, NA))
x
replace_na_with(x, 0L)

library(dplyr)
dat <- tibble(
  ones = c(NA,1,1),
  twos = c(NA,NA, 2),
  threes = c(NA, NA, NA)
)

dat

dat \%>\%
  mutate(
    ones = replace_na_with(ones, 0),
    twos = replace_na_with(twos, -2),
    threes = replace_na_with(threes, -3)
  )

dat \%>\%
  mutate(
    across(
      everything(),
      \(x) replace_na_with(x, -99)
    )
  )

}
