% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-to-na.R
\name{replace_to_na}
\alias{replace_to_na}
\title{Replace to NA}
\usage{
replace_to_na(x, ...)
}
\arguments{
\item{x}{the dataframe or vector containing the strings to replace}

\item{...}{the character strings}
}
\value{
a dataframe or vector
}
\description{
Replace one or more character string(s) with a NA
}
\examples{
replace_to_na(breizh, "not_available", "empty")
replace_to_na(breizh$Taille, "not_available", "empty")

}
