% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped-replace-to-na.R
\name{replace_to_na_at}
\alias{replace_to_na_at}
\title{Replace specified variables with NA where a certain condition is met}
\usage{
replace_to_na_at(data, .vars, .funs)
}
\arguments{
\item{data}{dataframe}

\item{.vars}{The variables to refer to}

\item{.funs}{the condition required to be TRUE to set NA - the conditions are
referred to as ~.x {condition} - e.g., "~.x < 20" means "where the
value is less than 20."}
}
\value{
a dataframe
}
\description{
Replace specified variables with NA where a certain condition is met
}
\examples{

dat_ms <- tibble::tribble(~x,  ~y,    ~z,
                          1,   "A",   -100,
                          3,   "N/A", -99,
                          NA,  NA,    -98,
                          -99, "E",   -101,
                          -98, "F",   -1)

dat_ms

replace_to_na_at(data = dat_ms,
                 .vars = "x",
                 .funs = ~.x == -99)

replace_to_na_at(data = dat_ms,
                 .vars = c("x","z"),
                 .funs = ~.x == -99)


}
