% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped-replace-to-na.R
\name{replace_to_na_if}
\alias{replace_to_na_if}
\title{Replace values with NA based on some condition, for variables that meet some predicate}
\usage{
replace_to_na_if(data, .predicate, .funs)
}
\arguments{
\item{data}{dataframe}

\item{.predicate}{a predicate function to be applied to the columns or a logical vector}

\item{.funs}{the condition required to be TRUE to set NA - the conditions are
referred to as ~.x {condition} - e.g., "~.x < 20" means "where the
value is less than 20."}
}
\value{
a dataframe
}
\description{
Replace values with NA based on some condition, for variables that meet some predicate
}
