% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped-replace-to-na.R
\name{replace_to_na_where}
\alias{replace_to_na_where}
\title{Replace all values with NA where a certain condition is met}
\usage{
replace_to_na_where(data, .p)
}
\arguments{
\item{data}{A dataframe}

\item{.p}{the condition required to be TRUE to set NA - the conditions are
referred to as ~.x {condition} - e.g., "~.x < 20" means "where the
value is less than 20."}
}
\description{
This function takes a dataframe and replaces all values that meet the
condition specified as an NA value, following a special syntax.
}
\examples{
# Replace all values less than 20
replace_to_na_where(airqaulity, ~ .x < 20)

# where works with functions
replace_to_na_where(airquality, ~ sqrt(.x) < 5)

}
