% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_missing.R
\name{row_missing}
\alias{row_missing}
\title{row_missing}
\usage{
row_missing(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
Return a tidy dataframe that contains the number of missings in each column
}
\examples{

row_missing(airquality)

}

