% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-n-prop-miss.R
\name{set-prop-n-miss}
\alias{set-prop-n-miss}
\alias{set_prop_miss}
\alias{set_n_miss}
\title{Set a proportion or number of missing values}
\usage{
set_prop_miss(x, prop = 0.1)

set_n_miss(x, n = 1)
}
\arguments{
\item{x}{vector of values to set missing}

\item{prop}{proportion of values between 0 and 1 to set as missing}

\item{n}{number of values to set missing}
}
\value{
vector with missing values added
}
\description{
Set a proportion or number of missing values
}
\examples{
vec <- rnorm(5)
set_prop_miss(vec, 0.2)
set_prop_miss(vec, 0.4)
set_n_miss(vec, 1)
set_n_miss(vec, 4)
}
