% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_cat.R
\name{shadow_cat}
\alias{shadow_cat}
\title{shadow_cat}
\usage{
shadow_cat(x)
}
\arguments{
\item{x}{a dataframe coming from `interaction`}
}
\description{
shadow_cat is a utility function that rearranges factor levels
}
\details{
\code{shadow_cat} reorganizes
}

