% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_df.R
\name{shadow_df}
\alias{shadow_df}
\title{shadow_df}
\usage{
shadow_df(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
\code{shadow_df} creates a shadow matrix/data frame of class 'tbl_df' that denotes whether a given cell is missing or not - if a value is missing, it is denoted as TRUE
}

