% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-shifters.R
\name{shadow_shift.numeric}
\alias{shadow_shift.numeric}
\title{Shift (impute) numeric values for graphical exploration}
\usage{
\method{shadow_shift}{numeric}(x, prop_below = 0.1, jitter = 0.05,
  seed_shift = 2017 - 7 - 1 - 1850, ...)
}
\arguments{
\item{x}{a variable of interest to shift}

\item{prop_below}{the degree to shift the values. default is}

\item{jitter}{the amount of jitter to add. default is 0.05}

\item{seed_shift}{a random seed to set, if you like}

\item{...}{extra arguments to pass}
}
\description{
Shift (impute) numeric values for graphical exploration
}
