% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_missing_histogram.R
\name{stat_missing_histogram}
\alias{stat_missing_histogram}
\title{stat_missing_histogram}
\usage{
stat_missing_histogram(mapping = NULL, data = NULL, geom = "histogram",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{?}{unknown parameters...}
}
\description{
stat_missing_histogram adds a geometry for displaying missingness to
geom_histogram
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
Warning message if na.rm = T is supplied.
}

