% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_missing_point.R
\name{stat_missing_point}
\alias{stat_missing_point}
\title{stat_missing_point}
\usage{
stat_missing_point(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{?}{unknown parameters...}
}
\description{
stat_missing_point adds a geometry for displaying missingness to geom_point
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
Warning message if na.rm = T is supplied.
}

