% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-tidy-miss.R
\name{summary_missing_case}
\alias{summary_missing_case}
\title{Summarise the missingness in each case}
\usage{
summary_missing_case(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_case_summary instead.
}
\seealso{
miss_case_summary
}
