% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{summary_missing_var}
\alias{summary_missing_var}
\title{summary_missing_var}
\usage{
summary_missing_var(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a data_frame of the percent of missing data in each variable
}
\description{
summary_missing_var
}

