% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{table_missing_case}
\alias{table_missing_case}
\title{Tabulate missings in cases.}
\usage{
table_missing_case(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Provide a tidy table of the number of cases with 0, 1, 2, up to n, missing values and the proportion of the number of cases those cases make up
}
\examples{

table_missing_case(airquality)


}
