% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{table_missing_var}
\alias{table_missing_var}
\title{table_missing_var}
\usage{
table_missing_var(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a dataframe of the
}
\description{
provide a tidy table of the number of variables with 0, 1, 2, up to n, missing values and the proportion of the number of variablers those variables make up
}
\examples{

table_missing_var(airquality)


}

