% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{tally_case_missing}
\alias{tally_case_missing}
\title{tally_case_missing}
\usage{
tally_case_missing(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a dataframe of the
}
\description{
provide a tidy table of the number of cases with 0, 1, 2, up to n, missing values and the proportion of the number of cases those cases make up
}
\examples{



}

