% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{tally_var_missing}
\alias{tally_var_missing}
\title{tally_var_missing}
\usage{
tally_var_missing(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a dataframe
}
\description{
provide a data_frame of the number of variables with 0, 1, 2, up to n, missing values and the proportion of the number of variables those variables make up
}

