% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_if_dataframe}
\alias{test_if_dataframe}
\title{Test if input is a data.frame}
\usage{
test_if_dataframe(x)
}
\arguments{
\item{x}{object}
}
\value{
an error if input (x) is a data.frame
}
\description{
Test if input is a data.frame
}
\examples{
\dontrun{
# success
test_if_dataframe(airquality)
#fail
my_test <- matrix(10)
test_if_dataframe(my_test)
}

}
