% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_if_missing}
\alias{test_if_missing}
\title{Test if the input is Missing}
\usage{
test_if_missing(x)
}
\arguments{
\item{x}{object}
}
\value{
an error if input (x) is not specified
}
\description{
Test if the input is Missing
}
\examples{
\dontrun{
# success
my_test <- x
test_if_null(my_test)
#fail
test_if_missing()
}
}
