% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_if_null}
\alias{test_if_null}
\title{Test if the input is NULL}
\usage{
test_if_null(x)
}
\arguments{
\item{x}{object}
}
\value{
an error if input (x) is NULL
}
\description{
Test if the input is NULL
}
\examples{
\dontrun{
# success
test_if_null(airquality)
#fail
my_test <- NULL
test_if_null(my_test)
}
}
