% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_if_shadow}
\alias{test_if_shadow}
\title{Test if input is a shadow}
\usage{
test_if_shadow(x)
}
\arguments{
\item{x}{object}
}
\value{
an error if input (x) is a shadow
}
\description{
Test if input is a shadow
}
\examples{
\dontrun{
# success
aq_shadow <- bind_shadow(airquality)
test_if_shadow(aq_shadow)
#fail
test_if_shadow(airquality)
}

}
