% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{what_levels}
\alias{what_levels}
\title{check the levels of many things}
\usage{
what_levels(x)
}
\arguments{
\item{x}{data.frame, usually}
}
\value{
a list containing the levels of everything
}
\description{
this function is used internally to check what the levels are of the dataframe.
}
