% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{where_na}
\alias{where_na}
\title{Which rows and cols contain missings?}
\usage{
where_na(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
integer locations of where the missing values are
}
\description{
Internal function that is short for `which(is.na(x))`. Creates integer
locations of missing values in a dataframe. May be used in future `impl_df`
class.
}
\examples{

narnia:::which_na(airquality)

}
\seealso{
which_na
}
