% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{which_are_shade}
\alias{which_are_shade}
\title{Which variables are shades?}
\usage{
which_are_shade(.tbl)
}
\arguments{
\item{.tbl}{a data.frame or tbl}
}
\value{
numeric - which column numbers contain shade information
}
\description{
This function tells us which variables contain shade information
}
\examples{

df_shadow <- bind_shadow(airquality)

which_are_shade(df_shadow)

}
