% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{which_are_shadow}
\alias{which_are_shadow}
\title{Which variables are shadows?}
\usage{
which_are_shadow(.tbl)
}
\arguments{
\item{.tbl}{a data.frame or tbl}
}
\value{
numeric - which column numbers contain shadow information
}
\description{
This function tells us which variables contain shadow information
}
\examples{

df_shadow <- bind_shadow(airquality)

which_are_shadow(df_shadow)

}
