% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_missing.R
\name{which_missing}
\alias{which_missing}
\title{which_missing}
\usage{
which_missing(x)
}
\arguments{
\item{x}{}
}
\value{
a ggplot plot
}
\description{
\code{which_missing} (need a better name!) produces a set of rectangles that indicate whether there is a missing element in a column or not
}

