% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where-na.R
\name{which_na}
\alias{which_na}
\title{Which elements contain missings?}
\usage{
which_na(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
integer locations of missing values.
}
\description{
Equivalent to \code{which(is.na())} - returns integer locations of missing values.
}
\examples{

which_na(airquality)

}
\seealso{
\code{\link[=where_na]{where_na()}}
}
