% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where-na.R
\name{which_var_na}
\alias{which_var_na}
\title{Which variables contain missing values?}
\usage{
which_var_na(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
character vector of variable names
}
\description{
It can be helpful when writing other functions to just return the names
of the variables that contain missing values.
}
\examples{

which_var_na(airquality)

which_var_na(iris)

}
