% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_any_miss}
\alias{add_any_miss}
\title{Add a column that tells you if there are any missing values}
\usage{
add_any_miss(data, vars, label = "any_miss")
}
\arguments{
\item{data}{data.frame}

\item{vars}{quoted variables}

\item{label}{label for the column, defaults to "any_miss"}
}
\value{
data.frame with data and the column labelling whether that row (for
    those variables) has any missing values - indicated by "missing" and
    "complete".
}
\description{
Add a column that tells you if there are any missing values
}
\note{
perhaps there should be a general function for this
}
\examples{

add_any_miss(airquality, c("Ozone", "Solar.R"))

}
