% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{add_shadow}
\alias{add_shadow}
\title{Add a shadow column to dataframe}
\usage{
add_shadow(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{One or more unquoted variable names, separated by commas. These also
respect the dplyr verbs \code{starts_with}, \code{contains}, \code{ends_with}, etc.}
}
\value{
data.frame
}
\description{
As an alternative to \code{bind_shadow()}, you can add specific individual shadow
columns to a dataset. These also respect the dplyr verbs
\code{starts_with}, \code{contains}, \code{ends_with}, etc.
}
\examples{

airquality \%>\% add_shadow(Ozone)
airquality \%>\% add_shadow(Ozone, Solar.R)

}
\seealso{
\code{\link[=bind_shadow]{bind_shadow()}} \code{\link[=add_any_miss]{add_any_miss()}} \code{\link[=add_label_missings]{add_label_missings()}} \code{\link[=add_label_shadow]{add_label_shadow()}} \code{\link[=add_miss_cluster]{add_miss_cluster()}} \code{\link[=add_n_miss]{add_n_miss()}} \code{\link[=add_prop_miss]{add_prop_miss()}} \code{\link[=add_shadow_shift]{add_shadow_shift()}} \code{\link[=cast_shadow]{cast_shadow()}}
}
