% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{add_shadow}
\alias{add_shadow}
\title{Add a column of the shadows to the dataframe}
\usage{
add_shadow(data, vars)
}
\arguments{
\item{data}{data.frame}

\item{vars}{quoted character string}
}
\value{
data.frame
}
\description{
Add a column of the shadows to the dataframe
}
\examples{

airquality \%>\% add_shadow(c("Ozone", "Solar.R"))

}
