% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{add_shadow}
\alias{add_shadow}
\title{Add a shadow column to dataframe}
\usage{
add_shadow(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{One or more unquoted expressions separated by commas. These also
respect the dplyr verbs \code{starts_with}, \code{contains}, \code{ends_with}, etc.}
}
\value{
data.frame
}
\description{
As an alternative to \code{bind_shadow()}, you can add specific individual shadow
columns to a dataset. These also respect the dplyr verbs
\code{starts_with}, \code{contains}, \code{ends_with}, etc.
}
\examples{

airquality \%>\% add_shadow(Ozone)
airquality \%>\% add_shadow(Ozone, Solar.R)

}
