% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{any-na}
\alias{any-na}
\alias{any_na}
\title{Identify if there are any missing values}
\usage{
any_na(x)
}
\arguments{
\item{x}{an R object to be tested}
}
\description{
It is useful to search for any instances of missing values. By default
handles atomic vectors without a class and NULL. It calls any(is.na(x))
on objects with classes and for recursive = FALSE, on lists and
pairlists. This means that you can get FALSE when you search through a
list that contains missings. So, \code{any_na} is shorthand for
\code{anyNA(x, recursive = TRUE)}.
}
\examples{

# any_na() works recursively with lists
bag_o_lists <- list(A = 1:5,
                    B = c(NA, 5:8),
                    C = c("A","NA"))

anyNA(bag_o_lists)
any_na(bag_o_lists)

anyNA(airquality)
any_na(airquality)


}
