% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{as_shadow}
\alias{as_shadow}
\title{Create shadow data}
\usage{
as_shadow(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe with appended
}
\description{
Return a tibble that in shadow matrix form, where the variables are the same but have a suffix _NA attached to indicate their difference.
}
\examples{

as_shadow(airquality)

}
