% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-add-cols.R
\name{cast_shadow}
\alias{cast_shadow}
\title{Add a shadow column to a dataset}
\usage{
cast_shadow(data, vars)
}
\arguments{
\item{data}{data.frame or .tbl}

\item{vars}{quoted variablename}
}
\value{
.data with the added variable shifted named as `var_NA`
}
\description{
Shifting the values to make them easier to display
}
\examples{

airquality \%>\% cast_shadow("Ozone")

}
