% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-shadows.R
\name{cast_shadow_shift}
\alias{cast_shadow_shift}
\title{Add a shadow and a shadow_shift column to a dataset}
\usage{
cast_shadow_shift(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{One or more unquoted variable names separated by commas. These
respect the dplyr verbs \code{starts_with}, \code{contains}, \code{ends_with}, etc.}
}
\value{
data.frame with the shadow and shadow_shift vars
}
\description{
Shift the values and add a shadow column.  It also respects the dplyr
verbs \code{starts_with}, \code{contains}, \code{ends_with}, etc.
}
\examples{

airquality \%>\% cast_shadow_shift(Ozone,Temp)

airquality \%>\% cast_shadow_shift(dplyr::contains("o"))

}
\seealso{
\code{\link[=cast_shadow_shift]{cast_shadow_shift()}}, \code{\link[=cast_shadow_shift_label]{cast_shadow_shift_label()}} \code{\link[=bind_shadow]{bind_shadow()}} \code{\link[=add_any_miss]{add_any_miss()}} \code{\link[=add_label_missings]{add_label_missings()}} \code{\link[=add_label_shadow]{add_label_shadow()}} \code{\link[=add_miss_cluster]{add_miss_cluster()}} \code{\link[=add_prop_miss]{add_prop_miss()}} \code{\link[=add_shadow_shift]{add_shadow_shift()}}
}
