% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{cast_shadow_shift}
\alias{cast_shadow_shift}
\title{Add a shadow and a shadow_shift column to a dataset}
\usage{
cast_shadow_shift(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{One or more unquoted expressions separated by commas. These also
respect the dplyr verbs "starts_with", "contains", "ends_with", etc.}
}
\value{
data.frame with the shadow and shadow_shift vars
}
\description{
Shift the values and add the shadow
}
\examples{

airquality \%>\% cast_shadow_shift(Ozone)
airquality \%>\% cast_shadow_shift(Ozone,Temp)

airquality \%>\% cast_shadow_shift(dplyr::contains("o"))

}
