% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{gather_shadow}
\alias{gather_shadow}
\title{gather_shadow}
\usage{
gather_shadow(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe in long, or  "molten" format, containing information about the missings
}
\description{
gather exists as we want to include this extra metadata about the rows that have missing data, but I also wanted to include some extra information about the class of the data, in case we need to gather the data back into a wider, rather than long, format. Here it takes a function `visdat`, `visdat:::fingerprint`, which is currently not a particularly complex function.
}
\examples{

aq_shadow <- gather_shadow(airquality)

aq_shadow

}

