% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-case.R
\name{gg_miss_case}
\alias{gg_miss_case}
\title{Plot the number of missings per case (row)}
\usage{
gg_miss_case(x, group, order_cases = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{group}{bare variable name, if you want to create a faceted plot.}

\item{order_cases}{logical Order the rows by missingness (default is no - FALSE)}
}
\value{
a ggplot object depicting the number of missings in a given case.
}
\description{
This is a visual analogue to \code{miss_case_summary}. It draws a ggplot of the
number of missings in each case (row). A default minimal theme is used, which
can be customised as normal for ggplot.
}
\examples{

gg_miss_case(airquality)
library(ggplot2)
gg_miss_case(airquality) + labs(x = "Number of Cases")
gg_miss_case(airquality, order_cases = TRUE)
gg_miss_case(airquality, group = Month)
gg_miss_case(airquality, group = Month, order_cases = TRUE)

}
