% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-x.R
\name{gg_miss_case}
\alias{gg_miss_case}
\title{Plot the number of missings per case (row)}
\usage{
gg_miss_case(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object depicting the number of missings in a given case.
}
\description{
This function draws a ggplot of the number of missings in each row.
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

gg_miss_case(airquality)
library(ggplot2)
gg_miss_case(airquality) + labs(x = "Number of Cases")

}
