% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_case_cumsum}
\alias{gg_miss_case_cumsum}
\title{Lineplot of the cumulative sum of missing for cases}
\usage{
gg_miss_case_cumsum(x, breaks = 20)
}
\arguments{
\item{x}{a dataframe}

\item{breaks}{the breaks for the x axis default is 20}
}
\value{
a ggplot object depicting the number of missings
}
\description{
A lineplot with the cumulative sum of missing values, reading the rows from the top to bottom
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

gg_miss_case_cumsum(airquality)
library(ggplot2)
gg_miss_case_cumsum(x = breizh, breaks = 50) + theme_bw()
}
