% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_span}
\alias{gg_miss_span}
\title{Plot the number of missings in a given repeating span}
\usage{
gg_miss_span(data, var, span_size)
}
\arguments{
\item{data}{data.frame}

\item{var}{a bare unquoted variable name}

\item{span_size}{integer describing the length of the span to be explored}
}
\value{
ggplot2 object
}
\description{
This is a replacement function to
imputeTS::plotNA.distributionBar(tsNH4, breaksize = 100), which shows the
number of missings in a given span
}
\examples{

gg_miss_span(pedestrian, span_size = 100)
}
