% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_span}
\alias{gg_miss_span}
\title{Plot the number of missings in a given repeating span}
\usage{
gg_miss_span(data, var, span_every)
}
\arguments{
\item{data}{data.frame}

\item{var}{a bare unquoted variable name}

\item{span_every}{integer describing the length of the span to be explored}
}
\value{
ggplot2 object
}
\description{
This is a replacement function to
imputeTS::plotNA.distributionBar(tsNH4, breaksize = 100), which shows the
number of missings in a given span.
}
\details{
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

\dontrun{
miss_var_span(pedestrian, hourly_counts, span_every = 3000)
library(ggplot2)
gg_miss_span(pedestrian, hourly_counts, span_every = 3000) + theme_classic()
}
}
