% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_var}
\alias{gg_miss_var}
\title{Plot the number of missings for each variable}
\usage{
gg_miss_var(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object depicting the number of missings in a given column
}
\description{
This function draws a ggplot plot of the number of missings in each column, rearranged to show which variables have the most missing data.
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

gg_miss_var(airquality)
library(ggplot2)
gg_miss_var(airquality) + labs(y = "Look at all the missing ones")

}
