% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-x.R
\name{gg_miss_var}
\alias{gg_miss_var}
\title{Plot the number of missings for each variable}
\usage{
gg_miss_var(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object depicting the number of missings in a given column
}
\description{
This is a visual analogue to `miss_var_summary`. It draws a ggplot of the
  number of missings in each variable, ordered to show which variables have
  the most missing data. A default minimal rtheme is used, which can be
  customised as normal for ggplot.
}
\examples{

gg_miss_var(airquality)
library(ggplot2)
gg_miss_var(airquality) + labs(y = "Look at all the missing ones")

}
