% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-x.R
\name{gg_miss_var_cumsum}
\alias{gg_miss_var_cumsum}
\title{Plot of cumulative sum of missing value for each variable}
\usage{
gg_miss_var_cumsum(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a ggplot object
}
\description{
A plot showing the cumulative sum of missing values for each variable,
reading columns from the left to the right of the initial dataframe. A
default minimal theme is used, which can be customised as normal for ggplot.
}
\examples{

gg_miss_var_cumsum(airquality)
}
