% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_miss_var_cumsum}
\alias{gg_miss_var_cumsum}
\title{Lineplot of the cumulative sum of missing value for each variable}
\usage{
gg_miss_var_cumsum(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object depicting the number of missings
}
\description{
A lineplot with the cumulative sum of missing values, reading columns from the left to
the right of your dataframe.
The plot is a ggplot object with a basic customisation.
You can customise it the way you wish, just like classic ggplot.
}
\examples{

gg_miss_var_cumsum(airquality)
}
