% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-which.R
\name{gg_miss_which}
\alias{gg_miss_which}
\title{Plot which variables contain a missing value}
\usage{
gg_miss_which(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object of which variables contains missing values
}
\description{
This plot produces a set of rectangles indicating whether there is a missing
element in a column or not.  A default minimal theme is used, which can be
customised as normal for ggplot.
}
\examples{

gg_miss_which(airquality)
}
\seealso{
\code{\link[=geom_miss_point]{geom_miss_point()}} \code{\link[=gg_miss_case]{gg_miss_case()}} \code{\link[=gg_miss_case_cumsum]{gg_miss_case_cumsum()}} \code{\link[=gg_miss_fct]{gg_miss_fct()}} \code{\link[=gg_miss_span]{gg_miss_span()}} \code{\link[=gg_miss_var]{gg_miss_var()}} \code{\link[=gg_miss_var_cumsum]{gg_miss_var_cumsum()}} \code{\link[=gg_miss_which]{gg_miss_which()}}
}
