% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-which.R
\name{gg_miss_which}
\alias{gg_miss_which}
\title{Plot which variables contain a missing value}
\usage{
gg_miss_which(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot object
}
\description{
This plot produces a set of rectangles indicating whether there is a missing
element in a column or not.  A default minimal theme is used, which can be
customised as normal for ggplot.
}
\examples{

gg_miss_which(airquality)
library(ggplot2)

}
