% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing_point.R, R/stat_missing_histogram.R, R/stat_missing_point.R
\docType{data}
\name{GeomMissingPoint}
\alias{GeomMissingPoint}
\alias{StatMissingHistogram}
\alias{StatMissingPoint}
\title{stat_missing_histogram}
\arguments{
\item{?}{unknown parameters...}

\item{?}{unknown parameters...}
}
\description{
stat_missing_histogram adds a geometry for displaying missingness to
geom_histogram

stat_missing_point adds a geometry for displaying missingness to geom_point
}
\note{
Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
Warning message if na.rm = T is supplied.

Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
Warning message if na.rm = T is supplied.
}
\keyword{datasets}

