% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_missing.R
\docType{data}
\name{StatMissing}
\alias{StatMissing}
\title{stat_missing}
\arguments{
\item{?}{unknown parameters...}
}
\description{
stat_missing adds a geometry for displaying missingness.
}
\note{
At this stage stat_missing is limited to geom_point

Very first attempt at creating a geom that is compatible with ggplot2.
Data plotting works. Still todo:
manipulate the colour aes so that the colours and legend appear.
fix awful default point sizes.
Warning message if na.rm = T is supplied.
}
\keyword{datasets}

