% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_below.R
\name{impute_below}
\alias{impute_below}
\title{Impute data with values shifted 10\% below range.}
\usage{
impute_below(...)
}
\arguments{
\item{...}{extra arguments to pass - see \code{\link[=shadow_shift]{shadow_shift()}} for discussion on this.}
}
\description{
It can be useful in exploratory graphics to impute data outside the range of
the data. \code{impute_below} imputes all variables with missings to have
values 10% below the range for numeric values, and for character or factor
values adds a new string or label. It is powered by \code{shadow_shift}, so
please see the documentation for \code{\link[=shadow_shift]{shadow_shift()}} to full details on the
different implementations.
}
