% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-zero.R
\name{impute_zero}
\alias{impute_zero}
\title{Impute zero into a vector with missing values}
\usage{
impute_zero(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with a fixed values replaced
}
\description{
This can be useful if you are imputing specific values, however we would
generally recommend to impute using other model based approaches. See
\code{\link[=impute_lm]{impute_lm()}} and friends.
}
\examples{

vec <- rnorm(10)

vec[sample(1:10, 3)] <- NA

impute_zero(vec)
}
