% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{is_shadow}
\alias{is_shadow}
\title{Is this thing a shadow?}
\usage{
is_shadow(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
logical - single value. TRUE if contains a variable with a column ending in "_NA"
}
\description{
Does this thing contain a shadow variable?
}
\examples{

df_shadow <- bind_shadow(airquality)

is_shadow(df_shadow)

}
