% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/shadow-verifiers.R
\name{is_shadow}
\alias{is_shadow}
\alias{are_shadow}
\title{Test if input is or are shadow variables}
\usage{
is_shadow(x)

are_shadow(x)
}
\arguments{
\item{x}{a vector or data.frame}
}
\examples{

aq_sh <- as_shadow(airquality)
aq_bind <- bind_shadow(airquality)

is_shadow(aq_sh)
are_shadow(aq_sh)
is_shadow(aq_bind)
are_shadow(aq_bind)
=======
% Please edit documentation in R/shadows.R
\name{is_shadow}
\alias{is_shadow}
\title{Is this thing a shadow?}
\usage{
is_shadow(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
logical - single value. TRUE if contains a variable with a column ending in "_NA"
}
\description{
Does this thing contain a shadow variable?
}
\examples{

df_shadow <- bind_shadow(airquality)

is_shadow(df_shadow)
>>>>>>> 775fb6e6bcdb441237d0435ddcbf35f83d4f80e5

}
