% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_missings.R
\name{label_missings}
\alias{label_missings}
\title{Is there a missing value in the row of a dataframe?}
\usage{
label_missings(data)
}
\arguments{
\item{data}{a dataframe or set of vectors of the same length}
}
\value{
character vector of "Missing" and "Not Missing".
}
\description{
Creates a character vector describing whether there are missing or not missing values
}
\note{
could this be a nice window function? Like `data %>% mutate(add_is_missing(var1,var2))` ?
}
\examples{

label_missings(airquality)

library(dplyr)

airquality \%>\% mutate(is_missing = label_missings(airquality))

}
