% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{label_shadow}
\alias{label_shadow}
\title{Label shadow values as missing or not missing}
\usage{
label_shadow(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{extra variable to label}
}
\value{
"Missing" or "Not Missing"
}
\description{
Powers \code{add_label_shadow}. For the moment it is an internal function.
}
