% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{label_shadow}
\alias{label_shadow}
\title{Label shadow values as missing or not missing}
\usage{
label_shadow(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
"Missing" or "Not Missing"
}
\description{
This is used to power add_label_shadow. It may be exported later, but for the
moment this is an internal function.
}
