% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-complete-x-pct-prop.R
\name{miss-complete-var-pct}
\alias{miss-complete-var-pct}
\alias{miss_var_pct}
\alias{complete_var_pct}
\title{Percentage of variables containing missings or complete values}
\usage{
miss_var_pct(data)

complete_var_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing or complete
data
}
\description{
Deprecated. Please see \code{\link[=miss_var_pct]{miss_var_pct()}} and \code{\link[=complete_var_pct]{complete_var_pct()}}.
}
\seealso{
\code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
