% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-cumsum.R
\name{miss_case_cumsum}
\alias{miss_case_cumsum}
\title{Summarise the missingness in each case}
\usage{
miss_case_cumsum(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a tibble containing the number and percent of missing data in each
  case
}
\description{
Provide a data.frame containing each case (row), the number and percent of
  missing values in each case.
}
\examples{

miss_case_cumsum(airquality)

}
