% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R,
%   R/missing_data_tidiers.R
\name{percent_missing_case}
\alias{percent_missing_case}
\alias{miss_case_pct}
\title{Percentage of cases that contain a missing values.}
\usage{
percent_missing_case(data)

miss_case_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing value
}
\description{
Percentage of cases that contain a missing values.

Calculate the percentage of cases (rows) that contain a missing value.
}
\examples{

miss_case_pct(airquality)

}
