% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_case_pct}
\alias{miss_case_pct}
\title{Percentage of cases that contain a missing values.}
\usage{
miss_case_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing value
}
\description{
Calculate the percentage of cases (rows) that contain a missing value.
}
\examples{

miss_case_pct(airquality)

}
