% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R,
%   R/missing_data_tidiers.R
\name{table_missing_case}
\alias{table_missing_case}
\alias{miss_case_table}
\title{Tabulate missings in cases.}
\usage{
table_missing_case(data)

miss_case_table(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Tabulate missings in cases.

Provide a tidy table of the number of cases with 0, 1, 2, up to n, missing values and the proportion of the number of cases those cases make up
}
\examples{

miss_case_table(airquality)


}
