% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-table.R
\name{miss_case_table}
\alias{miss_case_table}
\title{Tabulate missings in cases.}
\usage{
miss_case_table(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Provide a tidy table of the number of cases with 0, 1, 2, up to n, missing
values and the proportion of the number of cases those cases make up.
}
\examples{

miss_case_table(airquality)

airquality \%>\%
  group_by(Month) \%>\%
  miss_case_table()

}
