% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_df_pct}
\alias{miss_df_pct}
\title{Proportion of missing data in a dataframe}
\usage{
miss_df_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the proportion of missing data in a dataframe
}
\description{
Calculate the percent of missing data in a dataframe.
}
\examples{

miss_df_pct(oceanbuoys)
miss_df_pct(riskfactors)

}
