% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-prop-pct-summary.R
\name{miss_prop_summary}
\alias{miss_prop_summary}
\title{Proportions of missings in data, variables, and cases.}
\usage{
miss_prop_summary(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Return missing data info about the dataframe, the variables, and the cases.
Specifically, returning how many elements in a dataframe contain a missing
value, how many elements in a variable contain a missing value, and how many
elements in a case contain a missing.
}
\examples{

miss_prop_summary(airquality)
library(dplyr)
airquality \%>\% group_by(Month) \%>\% miss_prop_summary()

}
\seealso{
\url{miss_case_pct} \url{miss_case_prop} \url{miss_case_summary} \url{miss_case_table} \url{miss_scan_count} \url{miss_summary} \url{miss_var_pct} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
