% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_summary}
\alias{miss_summary}
\title{Collate summary measures from narnia into one tibble}
\usage{
miss_summary(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
\code{miss_summary} performs all of the missing data helper summaries and puts them into a list. Perhaps in the future this can all be some sort of nested dataframe?
}
\examples{

s_miss <- miss_summary(breizh)
s_miss$miss_df_pct
s_miss$miss_case_table
# etc, etc, etc.

}
