% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{miss_summary}
\alias{miss_summary}
\title{Collate summary measures from narnia into one tibble}
\usage{
miss_summary(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
\code{summarise_missingness} performs all of the missing data helper summaries and puts them into a list. Perhaps in the future this can all be some sort of nested dataframe?
}
\examples{

s_miss <- miss_summary(airquality)
s_miss$percent_missing_df
s_miss$table_missing_case
# etc, etc, etc.

}
