% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_var_cumsum}
\alias{miss_var_cumsum}
\title{Cumsum the missingness in each variable}
\usage{
miss_var_cumsum(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a data_frame of the cumsum of missing data in each variable
}
\description{
Provide a data_frame containing the cumsum of number & percentage of missingness for each variable
}
\examples{

miss_var_cumsum(airquality)

}
