% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-pct-prop.R
\name{miss_var_pct}
\alias{miss_var_pct}
\title{Percentage of variables containing missings}
\usage{
miss_var_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing data
}
\description{
Calculate the percentage of variables that contain a single missing value.
}
\examples{

miss_var_pct(riskfactors)
miss_var_pct(oceanbuoys)

}
\seealso{
\url{miss_case_pct} \url{miss_case_prop} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
