% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{miss_var_pct}
\alias{miss_var_pct}
\title{Percentage of variables containing missings}
\usage{
miss_var_pct(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing data
}
\description{
Calculate the percentage of variables that contain a single missing value.
}
\examples{

library(naniar)

miss_var_pct(airquality)

}
