% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{miss_var_summary}
\alias{miss_var_summary}
\title{Summarise the missingness in each variable}
\usage{
miss_var_summary(data, order_output = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{order_output}{wether or not the output should be ordered by n_missing}
}
\value{
a data_frame of the percent of missing data in each variable
}
\description{
Provide a data_frame containing the variable names, the number of missing values, in each variable, and the percent of missing values in each variable.
}
\examples{

miss_var_summary(airquality)

}
