% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-summary.R
\name{miss_var_summary}
\alias{miss_var_summary}
\title{Summarise the missingness in each variable}
\usage{
miss_var_summary(data, order = FALSE, ...)
}
\arguments{
\item{data}{a data.frame}

\item{order}{a logical indicating whether to order the result by \code{n_miss}.
Defaults to TRUE. If FALSE, order of variables is the order input.}

\item{...}{extra arguments}
}
\value{
a tibble of the percent of missing data in each variable
}
\description{
Provide a summary for each variable of the number, percent missings, and
cumulative sum of missings of the order of the variables. By default,
it orders by the most missings in each variable.
}
\note{
\code{n_miss_cumsum} is calculated as the cumulative sum of missings in the
variables in the order that they are given in the data when entering
the function
}
\examples{

miss_var_summary(airquality)
miss_var_summary(oceanbuoys, order = TRUE)

# works with group_by from dplyr
library(dplyr)
airquality \%>\%
  group_by(Month) \%>\%
  miss_var_summary()

}
\seealso{
\code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table} \url{n_complete} \url{n_complete_row} \url{n_miss} \url{n_miss_row} \url{pct_complete} \url{pct_miss} \url{prop_complete} \url{prop_complete_row} \url{prop_miss}
}
