% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss-x-table.R
\name{miss_var_table}
\alias{miss_var_table}
\title{Tabulate the missings in the variables}
\usage{
miss_var_table(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Provide a tidy table of the number of variables with 0, 1, 2, up to n,
missing values and the proportion of the number of variables those
variables make up.
}
\examples{

miss_var_table(airquality)

library(dplyr)
airquality \%>\%
  group_by(Month) \%>\%
  miss_var_table()

}
\seealso{
\url{miss_case_pct} \url{miss_case_prop} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_pct} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table}
}
