% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R,
%   R/missing_data_tidiers.R
\name{table_missing_var}
\alias{table_missing_var}
\alias{miss_var_table}
\title{Tabulate the missings in the variables}
\usage{
table_missing_var(data)

miss_var_table(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Tabulate the missings in the variables

Provide a tidy table of the number of variables with 0, 1, 2, up to n, missing values and the proportion of the number of variablers those variables make up
}
\examples{

miss_var_table(airquality)


}
