% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_missing_histogram.R, R/geom_missing_point.R, R/naniar-ggproto.R, R/stat_missing_histogram.R, R/stat_missing_point.R
\docType{data}
\name{GeomMissingHistogram}
\alias{GeomMissingHistogram}
\alias{GeomMissingPoint}
\alias{StatMissingHistogram}
\alias{StatMissingPoint}
\alias{naniar-ggproto}
\title{naniar-ggroto}
\format{An object of class \code{GeomMissingHistogram} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomMissingHistogram

GeomMissingPoint

StatMissingHistogram

StatMissingPoint
}
\description{
These are the stat and geom overrides using ggproto from ggplot2 that make naniar work.
}
\keyword{datasets}

