% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naniar-package.R
\docType{package}
\name{naniar}
\alias{naniar-package}
\alias{naniar}
\title{naniar: Data Structures, Summaries, and Visualisations for Missing Data}
\description{
naniar is a package to make it easier to summarise and handle missing values
in R. It strives to do this in a way that is as consistent with tidyverse
principles as possible.  The work is fully discussed at Tierney & Cook (2023)
\url{doi:10.18637/jss.v105.i07}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njtierney/naniar}
  \item \url{https://naniar.njtierney.com/}
  \item Report bugs at \url{https://github.com/njtierney/naniar/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Miles McBain \email{miles.mcbain@gmail.com} (\href{https://orcid.org/0000-0003-2865-2548}{ORCID})
  \item Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild [contributor]
  \item Jim Hester \email{james.f.hester@gmail.com} [contributor]
  \item Luke Smith [contributor]
  \item Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID}) [contributor]
}

}
