% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naniar-package.R
\docType{package}
\name{naniar}
\alias{naniar}
\alias{naniar-package}
\title{naniar}
\description{
naniar is a package to make it easier to summarise and handle missing values
in R. It strives to do this in a way that is as consistent with tidyverse
principles as possible.
}
\seealso{
\url{add_any_miss} \url{add_label_missings} \url{add_label_shadow} \url{add_miss_cluster} \url{add_n_miss} \url{add_prop_miss} \url{add_shadow} \url{add_shadow_shift} \url{as_shadow} \url{bind_shadow} \url{cast_shadow} \url{cast_shadow_shift} \url{cast_shadow_shift_label} \url{draw_key_missing_point} \url{gather_shadow} \url{geom_miss_point} \url{gg_miss_case} \url{gg_miss_case_cumsum} \url{gg_miss_fct} \url{gg_miss_span} \url{gg_miss_var} \url{gg_miss_var_cumsum} \url{gg_miss_which} \url{label_miss_1d} \url{label_miss_2d} \url{label_missings} \code{\link[=pct_miss_case]{pct_miss_case()}} \code{\link[=prop_miss_case]{prop_miss_case()}} \code{\link[=pct_miss_var]{pct_miss_var()}} \code{\link[=prop_miss_var]{prop_miss_var()}} \code{\link[=pct_complete_case]{pct_complete_case()}} \code{\link[=prop_complete_case]{prop_complete_case()}} \code{\link[=pct_complete_var]{pct_complete_var()}} \code{\link[=prop_complete_var]{prop_complete_var()}} \code{\link[=miss_prop_summary]{miss_prop_summary()}} \url{miss_case_summary} \url{miss_case_table} \url{miss_summary} \url{miss_var_prop} \url{miss_var_run} \url{miss_var_span} \url{miss_var_summary} \url{miss_var_table} \url{n_complete} \url{n_complete_row} \url{n_miss} \url{n_miss_row} \url{pct_complete} \url{pct_miss} \url{prop_complete} \url{prop_complete_row} \url{prop_miss} \url{prop_miss_row} \url{replace_to_na} \url{replace_with_na} \url{replace_with_na_all} \url{replace_with_na_at} \url{replace_with_na_if} \url{shadow_shift} \url{stat_miss_point} \url{vis_miss} \url{where_na}
}
