% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{percent_missing_case}
\alias{percent_missing_case}
\title{percent_missing_case}
\usage{
percent_missing_case(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
numeric the percentage of cases that contain a missing value
}
\description{
calculate percent of missing data in each case (row)
}
\examples{

library(ggmissing)
percent_missing_case(airquality)

}

