% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R
\name{percent_missing_df}
\alias{percent_missing_df}
\title{Percentage of missing data in a dataframe}
\usage{
percent_missing_df(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_df_pct instead.
}
\seealso{
miss_df_pct
}
