% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{percent_missing_var}
\alias{percent_missing_var}
\title{Percentage of variables containing missings}
\usage{
percent_missing_var(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
numeric the percent of variables that contain missing data
}
\description{
Calculate the percentage of variables that contain a single missing value.
}
\examples{

library(naniar)

percent_missing_var(airquality)

}

