% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop-miss-complete.R
\name{prop_complete}
\alias{prop_complete}
\title{Return the proportion of complete values}
\usage{
prop_complete(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
numeric proportion of complete values
}
\description{
The complement to `prop_miss`
}
\examples{

prop_complete(airquality)
prop_complete(airquality$Ozone)

}
