% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{prop_complete}
\alias{prop_complete}
\title{Return the proportion of complete values}
\usage{
prop_complete(x)
}
\arguments{
\item{x}{a vector}
}
\value{
numeric proprtion of complete values
}
\description{
The complement to \code{prop_miss}
}
\examples{

prop_complete(airquality)
prop_complete(airquality$Ozone)

}
