% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{prop_miss}
\alias{prop_miss}
\title{Return the proportion of missing values}
\usage{
prop_miss(x)
}
\arguments{
\item{x}{a vector}
}
\value{
numeric the proportion of missing values
}
\description{
substitute for \code{mean(is.na(data))}
}
\examples{

prop_miss(airquality)
prop_miss(airquality$Ozone)

}
