% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-miss-family.R
\name{prop_na}
\alias{prop_na}
\title{Proporitons of missings in data, variables, and cases.}
\usage{
prop_na(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Return missing data info about the dataframe, the variables, and the cases. Specifically, returning how many elements in a dataframe contain a missing value, how many elements in a variable contain a missing value, and how many elements in a case contain a missing.
}
\examples{

prop_na(airquality)

}
