% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-miss-x.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vis_miss}
\alias{\%>\%}
\title{Objects exported from other packages}
\description{
Objects exported from other packages

Objects exported from other packages
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{visdat}{\code{\link[visdat]{vis_miss}}}
}}

