% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vis_miss}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{visdat}{\code{\link[visdat]{vis_miss}}}
}}
\keyword{internal}

