% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-to-na.R
\name{replace_to_na}
\alias{replace_to_na}
\title{Replace values with missings}
\usage{
replace_to_na(...)
}
\arguments{
\item{...}{additional arguments for methods.}
}
\value{
values replaced by NA
}
\description{
This function is Defunct, please see \code{\link[=replace_with_na]{replace_with_na()}}.
}
