% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped-replace-to-na.R
\name{replace_to_na_at}
\alias{replace_to_na_at}
\title{Replace specified variables with NA where a certain condition is met}
\usage{
replace_to_na_at(data, .vars, .p)
}
\arguments{
\item{data}{dataframe}

\item{.vars}{The variables to refer to}

\item{.p}{the condition required to be TRUE to set NA - the conditions are
referred to as ~.x {condition} - e.g., "~.x < 20" means "where the
value is less than 20."}
}
\value{
a dataframe
}
\description{
Replace specified variables with NA where a certain condition is met
}
\examples{
replace_to_na_at(airquality, .vars = c("Wind", "Ozone"), .p = ~ .x < 20)
}
