% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped-replace-to-na.R
\name{replace_to_na_where}
\alias{replace_to_na_where}
\title{Replace a value in a variable to NA according to some rule}
\usage{
replace_to_na_where(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{}
}
\description{
Replace a value in a variable to NA according to some rule
}
