% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_shadow_shifters.R
\name{shadow_cat}
\alias{shadow_cat}
\title{Categorise specific missings into missing  / not missing}
\usage{
shadow_cat(x)
}
\arguments{
\item{x}{dataframe coming from \code{interaction}}
}
\description{
\code{shadow_cat} is a window function that rearranges factor levels from \code{interaction}.
}

