% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadows.R
\name{shadow_long}
\alias{shadow_long}
\title{Reshape shadow data into a long format}
\usage{
shadow_long(shadow_data, vars_of_interest, only_main_vars = TRUE)
}
\arguments{
\item{shadow_data}{a data.frame}
}
\value{
data in the
}
\description{
Once data is in \code{nabular} form, where the shadow is bound to the data, it
can be useful to reshape it into a long format with the columns
}
\examples{

shadow_long(bind_shadow(airquality))

}
