% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-shifters.R
\name{shadow_shift.numeric}
\alias{shadow_shift.numeric}
\title{Shift (impute) numeric values for graphical exploration}
\usage{
\method{shadow_shift}{numeric}(x, seed_shift = 2017 - 7 - 1 - 1850,
  shift_prop = 0.1, jitter = 0.05, ...)
}
\arguments{
\item{shift_prop}{the degree to shift the values. default is}

\item{jitter}{the amount of jitter to add. deafult is 0.05}
}
\description{
Shift (impute) numeric values for graphical exploration
}
