% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_shift.R
\name{shadow_shift}
\alias{shadow_shift}
\title{shadow_shift}
\usage{
shadow_shift(x)
}
\arguments{
\item{x}{is a variable, must be continuous}
}
\description{
\code{shadow_shift} transforms missing values of a given variable
}

