% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-shifters.R
\name{shadow_shift}
\alias{shadow_shift}
\title{Shift missing values to facilite missing data exploration/visualisation}
\usage{
shadow_shift(x, ...)
}
\arguments{
\item{x}{a variable of interest to shift}

\item{...}{extra arguments to pass}
}
\description{
`shadow_shift` is a window function that transforms missing values to
    facilitate visualisation. This has different behaviour for different
    types of variables. For numeric variables, the values are shifted to 10%
    below the minimum value for a given variable plus some jittered noise,
    to separate repeated values, so that missing values can be visualised
    along with the rest of the data. There are different shifts that can
    take place in different kinds of variables, which are currently under
    development.
}
\examples{
airquality$Ozone
shadow_shift(airquality$Ozone)
library(dplyr)
airquality \%>\%
    mutate(Ozone_shift = shadow_shift(Ozone))

}
