% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_shifters.R
\name{shadow_shift}
\alias{shadow_shift}
\title{Shift missing values to 10\% below minimum value}
\usage{
shadow_shift(x)
}
\arguments{
\item{x}{is a variable, must be continuous}
}
\description{
shadow_shift is a window function that transforms missing values to be about 10% below the minimum value for a given variable, plus some jittered noise, to separate repeated values, so that missing values can be visualised along with the rest of the data
}
\examples{
airquality$Ozone
shadow_shift(airquality$Ozone)
library(dplyr)
airquality \%>\%
    mutate(Ozone_shift = shadow_shift(Ozone))

}

