% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-tidy-miss.R
\name{summarise_missingness}
\alias{summarise_missingness}
\title{Collate summary measures from narnia into one tibble}
\usage{
summarise_missingness(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_summary instead.
}
\seealso{
miss_summary
}
