% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{summarise_missingness}
\alias{summarise_missingness}
\title{summarise_missingness}
\usage{
summarise_missingness(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a list of
}
\description{
summarise_missingness performs all of the missing data helper summaries and puts them into a list. Perhaps in the future this can all be some sort of nested dataframe?
}

