% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{summary_missing_case}
\alias{summary_missing_case}
\title{summary_missing_case}
\usage{
summary_missing_case(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
a data_frame of the percent of missing data in each case
}
\description{
summary_missing_case
}
\examples{

summary_missing_case(airquality)

}

