% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{summary_missing_case}
\alias{summary_missing_case}
\title{Summarise the missingness in each case}
\usage{
summary_missing_case(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a data_frame of the percent of missing data in each case
}
\description{
Provide a data_frame containing each case (row), the number of missing values in each case,  and the percent of missing values in each case.
}
\examples{

summary_missing_case(airquality)

}

