% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{summary_missing_var}
\alias{summary_missing_var}
\title{Summarise the missingness in each variable}
\usage{
summary_missing_var(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a data_frame of the percent of missing data in each variable
}
\description{
Provide a data_frame of the number of variables with 0, 1, 2, up to n, missing values and the proportion of the number of variables those variables make up
}
\examples{

summary_missing_var(airquality)

}

