% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct_missing_data_tidiers.R
\name{summary_missing_var}
\alias{summary_missing_var}
\title{Summarise the missingness in each variable}
\usage{
summary_missing_var(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_var_summary instead.
}
\seealso{
miss_var_summary
}
