% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R
\name{table_missing_case}
\alias{table_missing_case}
\title{Tabulate missings in cases.}
\usage{
table_missing_case(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_case_table instead.
}
\seealso{
miss_case_table
}
