% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_tidiers.R
\name{table_missing_var}
\alias{table_missing_var}
\title{Tabulate the missings in the variables}
\usage{
table_missing_var(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Provide a tidy table of the number of variables with 0, 1, 2, up to n, missing values and the proportion of the number of variablers those variables make up
}
\examples{

table_missing_var(airquality)


}

