% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_missing_data_tidiers.R
\name{table_missing_var}
\alias{table_missing_var}
\title{Tabulate the missings in the variables}
\usage{
table_missing_var(...)
}
\arguments{
\item{...}{so that users can still receive meaningful error}
}
\description{
Deprecated as of version 0.0.6.9000. Please use miss_var_table instead.
}
\seealso{
miss_var_table
}
