% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where-na.R
\name{where_na}
\alias{where_na}
\title{Which rows and cols contain missings?}
\usage{
where_na(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
integers that describe the location of missing values
}
\description{
Internal function that is short for \code{which(is.na(x))}. Creates integer
locations of missing values in a dataframe. May be used in future \code{impl_df}
class.
}
\examples{

where_na(airquality)
where_na(oceanbuoys$sea_temp_c)

}
\seealso{
which_na
}
