% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_na}
\alias{which_na}
\title{Which elements contain missings?}
\usage{
which_na(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a matrix with columns "row" and "col", which refer to the row and
    column that identify the position of a missing value in a dataframe
}
\description{
Internal function that creates a matrix containing the location of missing
values in a dataframe. This may be used in the future `impl_df` class.
}
\examples{

narnia:::which_na(airquality)

}
\seealso{
where_na
}
